% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACETYL_NORM_DATA}
\alias{ACETYL_NORM_DATA}
\alias{ACETYL_HEART_NORM_DATA}
\alias{ACETYL_LIVER_NORM_DATA}
\title{Normalized protein acetylation data}
\format{
A data frame with acetylsites in rows (\code{feature_ID}) and samples in columns (\code{viallabel})

An object of class \code{data.frame} with 9750 rows and 54 columns.
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-untargeted/prot-ac/normalized-data/*med-mad-normalized-logratio.txt}
}
\usage{
ACETYL_HEART_NORM_DATA

ACETYL_LIVER_NORM_DATA
}
\description{
Median-MAD normalized protein acetylation (ACETYL) data used for visualization and differential analysis
}
\details{
Log2 TMT ratios to the common reference were used as quantitative values for all proteomics features
(proteins, phosphosites, acetylsite, and ubiquitylsites). Proteomics features not fully quantified in at least
two plexes within a tissue and non-rat contaminants were removed. Log2 TMT ratios were sample-normalized by
median-centering and mean absolute deviation scaling. Plex batch effects were removed using linear models
implemented by the \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.
}
\keyword{datasets}
