% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSAY_ABBREV}
\alias{ASSAY_ABBREV}
\title{Assay or "ome" abbreviations}
\format{
Unnamed character vector
}
\usage{
ASSAY_ABBREV
}
\description{
Assay/ome abbreviations used in tables and figures
}
\keyword{datasets}
