% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSAY_ABBREV_TO_CODE}
\alias{ASSAY_ABBREV_TO_CODE}
\title{Assay abbreviation-to-code mapping}
\format{
Named character vector
}
\usage{
ASSAY_ABBREV_TO_CODE
}
\description{
Values are codes and names are abbreviations \link{ASSAY_ABBREV}.
See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes} for more details.
}
\keyword{datasets}
