% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSAY_CODE_TO_ABBREV}
\alias{ASSAY_CODE_TO_ABBREV}
\title{Assay code-to-abbreviation mapping}
\format{
Named character vector
}
\usage{
ASSAY_CODE_TO_ABBREV
}
\description{
Values are abbreviations (\code{ASSAY_ABBREV}) and names are codes.
See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes} for more details.
}
\keyword{datasets}
