% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSAY_COLORS}
\alias{ASSAY_COLORS}
\title{Assay colors}
\format{
Named character vector
}
\usage{
ASSAY_COLORS
}
\description{
Assay colors used in figures. Values are hex codes and names are assay abbreviations (\code{ASSAY_ABBREV}).
}
\keyword{datasets}
