% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSAY_ORDER}
\alias{ASSAY_ORDER}
\title{Assay order}
\format{
Unnamed character vector
}
\usage{
ASSAY_ORDER
}
\description{
Biological order of \code{ASSAY_ABBREV} used in figures
}
\keyword{datasets}
