% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ATAC_DA}
\alias{ATAC_DA}
\title{Differential analysis of ATAC-seq data}
\format{
A data frame with 21 variables:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link{REPEATED_FEATURES} for details.}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{dataset}}{character, tissue code and Chemical Analysis Site that
generated the data. Note only ATAC data generated by Stanford are included
in this repository.}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{adj_p_value}}{double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{covariates}}{character, comma-separated list of adjustment variables or NA}
\item{\code{removed_samples}}{character, comma-separated list of vial labels excluded from differential analysis or NA}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{comparison_average_intensity_se}}{double, standard error of 'comparison_average_intensity'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{reference_average_intensity_se}}{double, standard error of 'reference_average_intensity'}
\item{\code{selection_fdr}}{double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.}
}
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
chromatin accessibility peak within each sex. One data frame per tissue.
}
\details{
Unfiltered ATAC differential analysis results are only available via download from Google Cloud Storage.
For example, \url{https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/ATAC_BAT_DA.rda}
is the file for brown adipose tissue (BAT) results. You can change the name of the file to specify other tissues including:
HEART, HIPPOC, KIDNEY, LIVER, LUNG, SKMGN (gastrocnemius skeletal muscle), and WATSC (subcutaneous white adipose tissue).
You can also use \code{\link[MotrpacRatTraining6mo:get_rdata_from_url]{MotrpacRatTraining6mo::get_rdata_from_url()}}
to download differential analysis results for ATAC and METHYL.
For more details about these files see the readme of this repository at
\url{https://github.com/MoTrPAC/MotrpacRatTraining6mo/blob/main/README.md}.

For training-regulated ATAC features at 5\% FDR, see \link{TRAINING_REGULATED_FEATURES}.

Reproduce our analysis with \code{\link[MotrpacRatTraining6mo:atac_timewise_da]{MotrpacRatTraining6mo::atac_timewise_da()}}
and \code{\link[MotrpacRatTraining6mo:atac_training_da]{MotrpacRatTraining6mo::atac_training_da()}}.
}
