% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATAC_META}
\alias{ATAC_META}
\title{ATAC-seq metadata and QC}
\format{
A data frame with 520 rows and 106 variables:
\describe{
\item{\code{viallabel}}{double, sample identifier}
\item{\code{general.description}}{character, pipeline workflow description}
\item{\code{replicate}}{character, replicate in the pipeline workflow}
\item{\code{general.date}}{double, date the pipeline workflow was run}
\item{\code{general.title}}{character, pipeline workflow title}
\item{\code{general.pipeline_ver}}{character, ENCODE ATAC-seq pipeline version}
\item{\code{general.pipeline_type}}{character, "atac"}
\item{\code{general.genome}}{character, reference genome}
\item{\code{general.aligner}}{character, read aligner}
\item{\code{general.peak_caller}}{character, peak caller}
\item{\code{general.seq_endedness.paired_end}}{logical, are the reads paired-ended}
\item{\code{replication.num_peaks.num_peaks}}{integer, number of replication peaks (max 300000)}
\item{\code{peak_stat.peak_region_size.min_size}}{integer, minimum peak width}
\item{\code{peak_stat.peak_region_size.25_pct}}{integer, 25th percentile of peak width}
\item{\code{peak_stat.peak_region_size.50_pct}}{integer, 50th percentile of peak width}
\item{\code{peak_stat.peak_region_size.75_pct}}{integer, 75th percentile of peak width}
\item{\code{peak_stat.peak_region_size.max_size}}{integer, max peak width}
\item{\code{peak_stat.peak_region_size.mean}}{double, mean peak width}
\item{\code{peak_enrich.frac_reads_in_peaks.macs2.frip}}{double, fraction of reads in MACS2 peaks}
\item{\code{align.samstat.total_reads}}{integer, total number of alignments, including multimappers}
\item{\code{align.samstat.mapped_reads}}{integer, total number of mapped reads}
\item{\code{align.samstat.pct_mapped_reads}}{double, percent of reads that mapped}
\item{\code{align.samstat.paired_reads}}{integer, number of paired reads}
\item{\code{align.samstat.read1}}{integer, number of read 1 reads}
\item{\code{align.samstat.read2}}{integer, number of read 2 reads}
\item{\code{align.samstat.properly_paired_reads}}{integer, number of properly paired reads}
\item{\code{align.samstat.pct_properly_paired_reads}}{double, percent of reads that were properly paired}
\item{\code{align.samstat.with_itself}}{integer, number of reads paired with its pair}
\item{\code{align.samstat.singletons}}{integer, number of singleton reads}
\item{\code{align.samstat.pct_singletons}}{double, percent of reads that were singleton}
\item{\code{align.samstat.diff_chroms}}{integer}
\item{\code{align.dup.paired_reads}}{integer, number of paired reads for duplication step}
\item{\code{align.dup.paired_duplicate_reads}}{integer, number of duplicate paired reads}
\item{\code{align.dup.paired_optical_duplicate_reads}}{integer, number of optical duplicate paired reads}
\item{\code{align.dup.pct_duplicate_reads}}{double, percent of reads that are duplicate}
\item{\code{align.frac_mito.non_mito_reads}}{integer, percent of reads that align to non-mitochondrial DNA}
\item{\code{align.frac_mito.mito_reads}}{integer, number of reads that align to mitochondrial DNA}
\item{\code{align.frac_mito.frac_mito_reads}}{double, fraction of reads that align to mitochondrial DNA}
\item{\code{align.nodup_samstat.total_reads}}{integer, number of reads after applying all filters}
\item{\code{align.nodup_samstat.mapped_reads}}{integer, number of mapped reads after applying all filters}
\item{\code{align.nodup_samstat.paired_reads}}{integer, number of paired reads after applying all filters}
\item{\code{align.nodup_samstat.read1}}{integer, number of read 1 reads after applying all filters}
\item{\code{align.nodup_samstat.read2}}{integer, number of read 2 reads after applying all filters}
\item{\code{align.nodup_samstat.properly_paired_reads}}{integer, number of properly paired reads after applying all filters}
\item{\code{align.nodup_samstat.with_itself}}{integer, number of reads paired with its pair after applying all filters}
\item{\code{align.frag_len_stat.frac_reads_in_nfr}}{double, fraction of reads in nucelosome-free region. Should be a value greater than 0.4.}
\item{\code{align.frag_len_stat.frac_reads_in_nfr_qc_pass}}{logical, does \code{align.frag_len_stat.frac_reads_in_nfr} pass the cutoff?}
\item{\code{align.frag_len_stat.frac_reads_in_nfr_qc_reason}}{character, reason for \code{align.frag_len_stat.frac_reads_in_nfr_qc_pass}}
\item{\code{align.frag_len_stat.nfr_over_mono_nuc_reads}}{double, reads in nucleosome-free-region
versus reads in mononucleosomal peak. Should be a value greater than 2.5.}
\item{\code{align.frag_len_stat.nfr_over_mono_nuc_reads_qc_pass}}{logical, does \code{align.frag_len_stat.frac_reads_in_nfr_qc_pass} pass the cutoff?}
\item{\code{align.frag_len_stat.nfr_over_mono_nuc_reads_qc_reason}}{character, reason for \code{align.frag_len_stat.nfr_over_mono_nuc_reads_qc_pass}}
\item{\code{align.frag_len_stat.nfr_peak_exists}}{logical, does a nucleosome-free peak exist?}
\item{\code{align.frag_len_stat.mono_nuc_peak_exists}}{logical, does a mono-nucleosomal peak exist?}
\item{\code{align.frag_len_stat.di_nuc_peak_exists}}{logical, does a di-nucleosomal peak exist?}
\item{\code{lib_complexity.lib_complexity.total_fragments}}{integer, total number of fragments}
\item{\code{lib_complexity.lib_complexity.distinct_fragments}}{integer, number of distinct fragments}
\item{\code{lib_complexity.lib_complexity.positions_with_one_read}}{integer, number of positions with one read}
\item{\code{lib_complexity.lib_complexity.NRF}}{double, non-reduandant fraction. Measure of library complexity. Ideally >0.9}
\item{\code{lib_complexity.lib_complexity.PBC1}}{double, PCR bottlenecking coefficient 1. Measure of library complexity. Ideally >0.9}
\item{\code{lib_complexity.lib_complexity.PBC2}}{double PCR bottlenecking coefficient 2. Measure of library complexity. Ideally >3}
\item{\code{align_enrich.tss_enrich.tss_enrich}}{double, transcription start site enrichment of peaks}
\item{\code{2D_barcode}}{double, sample barcode}
\item{\code{Tissue}}{character, tissue description}
\item{\code{Species}}{character, species}
\item{\code{Sample_category}}{character, study sample ("study") or reference standard ("ref)}
\item{\code{GET_site}}{character, which Genomics, Epigenomics, and Transcriptomics (GET)
site performed the assay, "Stanford" or "MSSM" (Icahn School of Medicine at Mount Sinai)}
\item{\code{Sample_batch}}{integer, numeric batch number for batch in which this sample was manually processed}
\item{\code{Lib_adapter_1}}{character, Adapter sequence for read 1}
\item{\code{Lib_adapter_2}}{character, Adapter sequence for read 2}
\item{\code{Lib_index_1}}{character, i7 index}
\item{\code{Lib_index_2}}{character, i5 index}
\item{\code{Nuclei_extr_date}}{character, nuclei extraction date}
\item{\code{Nuclei_extr_count}}{integer, nuclei count}
\item{\code{Nuclei_tagmentation}}{integer, number of nuclei used in each tagmentation reaction}
\item{\code{Tagmentation_date}}{character, tagmentation date, MM/DD/YYYY format}
\item{\code{Tagmentation_enzyme_cat}}{integer, catalog number of tagmentation enzyme TDE1 (Tn5)}
\item{\code{Tagmentation_enzyme_lot}}{integer, lot number of tagmentation enzyme TDE1 (Tn5)}
\item{\code{Tagmentation_buffer_cat}}{integer, catalog number of tagmentation buffer}
\item{\code{Tagmentation_buffer_lot}}{integer, lot number of tagmentation buffer}
\item{\code{Tagmentation_reaction_vol}}{integer, volume of tagmentation (uL)}
\item{\code{Tagmentation_purification_kit}}{character, purification kit}
\item{\code{Tagmentation_purified_DNA_vol}}{double, volume of purified DNA (uL)}
\item{\code{PCR_date}}{character, PCR date, MM/DD/YYYY format}
\item{\code{PCR_cycle_nr}}{integer, number of PCR cycles}
\item{\code{PCR_purification_beads_ul}}{character, volume of SPRIselect beads for lower size selection}
\item{\code{Lib_DNA_conc}}{double, DNA concentration for the library (ng/uL)}
\item{\code{Lib_DNA_molarity}}{double, DNA molarity of library (nM)}
\item{\code{Lib_frag_size}}{integer, average library fragment size}
\item{\code{Lib_BA_quality}}{integer, visual inspection of the library quality with the Bioanalyzer track (1=good, 0=bad)}
\item{\code{Seq_DNA_molarity}}{double, DNA molarity for sequencing (nM)}
\item{\code{Seq_platform}}{character, sequencing platform}
\item{\code{Seq_date}}{integer, sequencing date, YYMMDD format}
\item{\code{Seq_machine_ID}}{character, serial number of the sequencer}
\item{\code{Seq_flowcell_ID}}{character, flow cell ID}
\item{\code{Seq_flowcell_run}}{integer, flow cell run}
\item{\code{Seq_flowcell_lane}}{character, flow cell lane}
\item{\code{Seq_flowcell_type}}{character, flow cell type, e.g., S4}
\item{\code{Seq_length}}{double, read length}
\item{\code{Seq_end_type}}{integer, 1=single-end, 2=paired-end}
\item{\code{total_primary_alignments}}{integer, number of primary alignments}
\item{\code{pct_chrX}}{double, number of reads mapped to chromosome X}
\item{\code{pct_chrY}}{double, number of reads mapped to chromosome Y}
\item{\code{pct_chrM}}{double, number of reads mapped to chromosome M}
\item{\code{pct_auto}}{double, number of reads mapped to autosomal chromosomes}
\item{\code{pct_contig}}{double, number of reads mapped to contigs}
\item{\code{Seq_batch}}{character, unique identifier for sequencing batch}
}
}
\source{
\url{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/results/epigenomics/qa-qc/motrpac_pass1b-06_epigen-atac-seq_qa-qc-metrics.csv}
}
\usage{
ATAC_META
}
\description{
ATAC-seq experimental and quantification QC metrics for chromatin accessibility (ATAC) data
}
\details{
The \href{https://github.com/ENCODE-DCC/atac-seq-pipeline/tree/v1.7.0}{ENCODE ATAC-seq pipeline v1.7.0} was used to quantify ATAC-seq data.
Columns with a period are QC metrics from this pipeline. Note that the ENCODE pipeline reports alignments per paired-end read,
so \code{align.samstat.total_reads} reports the number of paired-end reads that align, which corresponds to twice the number of sequenced fragments.
}
\keyword{datasets}
