% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATAC_SAMPLE_DATA}
\alias{ATAC_SAMPLE_DATA}
\title{ATAC-seq normalized sample-level data for training-regulated features}
\format{
A data frame with 2247 rows and 51 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/epigenomics/epigen-atac-seq/normalized-data/*quant-norm*}
}
\usage{
ATAC_SAMPLE_DATA
}
\description{
\code{\link[limma:voom]{limma::voom()}} quantile-normalized ATAC-seq counts for training-regulated features only.
See the data release on \url{https://motrpac-data.org/} for sample-level data for all features.
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
