% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRAPH_COMPONENTS}
\alias{GRAPH_COMPONENTS}
\title{Graph components}
\format{
List of lists:
\describe{
\item{\code{edge_sets}}{named list, where the name is the label for the edge in the
format \code{[fromNode]---[toNode]} and the value is a vector of features that belong to that edge}
\item{\code{node_sets}}{named list, where the name is the label for the node in the
format \code{[week]w_F[state]_M[state]} and the value is a vector of features that belong to that edge}
}
}
\usage{
GRAPH_COMPONENTS
}
\description{
The feature sets selected using the \code{repfdr} results. Each
value is a named list, where the name is the label for the edge or node,
and the members are all of the features that belong to that label,
in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[feature_ID]}.
}
\details{
Nodes are named in the format \code{[week]w_F[state]_M[state]}, where
\code{[week]} is one of 1, 2, 4, or 8 to specify the training point, \code{[state]}
takes a value of 0 (null), 1 (up-regulated), or -1 (down-regulated). Altogether,
then name of a node specifies the state of a feature in each sex at a specific
time point. For example, the node \code{1w_F-1_M-1} specifies all features that
are down-regulated in females (\code{F-1}) and males (\code{M-1}) at the 1-week
training time point (\code{1w}).
}
\keyword{datasets}
