% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRAPH_PW_ENRICH}
\alias{GRAPH_PW_ENRICH}
\title{Graph pathway enrichment results}
\format{
A data frame with 156906 rows and 22 variables:
\describe{
\item{\code{query}}{character, not used, carried over from \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{significant}}{logical, not used, carried over from \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{term_size}}{double, effective pathway size from \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{query_size}}{integer, size of input, i.e. list of Ensembl genes associated with differential features}
\item{\code{intersection_size}}{double, size of the intersection between the input and the pathway members}
\item{\code{precision}}{double, the proportion of genes in the input list that are annotated to the function (defined as \code{intersection_size/query_size})}
\item{\code{recall}}{double, the proportion of functionally annotated genes that the query recovers (defined as \code{intersection_size/term_size})}
\item{\code{term_id}}{character, pathway term ID}
\item{\code{source}}{character, database corresponding to the pathway, one of: "KEGG", "REAC"}
\item{\code{term_name}}{character, pathway name}
\item{\code{effective_domain_size}}{integer, size of the custom background Ensembl gene set}
\item{\code{source_order}}{integer, not used, carried over from \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{parents}}{list, pathway parent(s)}
\item{\code{evidence_codes}}{character, not used, carried over from \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{intersection}}{character, intersection between input and pathway (Ensembl IDs). NA for metabolomics enrichments}
\item{\code{gost_adj_p_value}}{double, BH-adjusted p-value returned by \code{\link[gprofiler2:gost]{gprofiler2::gost()}},
ignored because p-values are only adjusted within each tissue/ome/cluster combination. Use the \code{adj_p_value} column instead.}
\item{\code{computed_p_value}}{double, nominal hypergeometric p-value, computed from the \code{\link[gprofiler2:gost]{gprofiler2::gost()}} output}
\item{\code{cluster}}{character, graphical cluster (node, edge, or path) name}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}. Note that VENACV, OVARY, TESTES,
were not included in the graphical representation of differential
features due to missing groups (e.g., females trained for 1 week).}
\item{\code{ome}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{kegg_id}}{character, pathway ID returned from \code{\link[FELLA:enrich-funs]{FELLA::enrich()}}}
\item{\code{adj_p_value}}{double, IHW FDR, calculated using \code{\link[IHW:ihw.default]{IHW::ihw()}} with \code{tissue} as a covariate}
\item{\code{graphical_cluster}}{character, \code{cluster} column with tissue prefix removed}
}
}
\usage{
GRAPH_PW_ENRICH
}
\description{
Pathway enrichment results for graphical clusters (nodes, edges, and paths) of interest
}
\details{
All non-metabolite training-regulated features (5\% FDR) were mapped
to Ensembl gene symbols using \link{FEATURE_TO_GENE}. Training-regulated metabolites
were mapped to KEGG IDs. For each graphical cluster of interest (i.e.,
the ten largest paths, two largest nodes, and two largest single edges with
at least 20 features in each tissue, as well as all 8-week nodes),
we performed pathway enrichment analysis separately for the Ensembl genes
(or KEGG IDs for metabolites) associated with differential features in each ome.

For gene-centric omes (i.e., all but metabolomics)
we performed enrichment analysis of KEGG and REACTOME rat pathways (organism "rnorvegicus")
using \code{\link[gprofiler2:gost]{gprofiler2::gost()}} with custom backgrounds defined by \link{GENE_UNIVERSES}.
Only pathways with at least 10 and up to 200 members were tested. Because \code{\link[gprofiler2:gost]{gprofiler2::gost()}}
only returns adjusted p-values, we recalculated nominal p-values using a one-tailed hypergeometric test,
which is consistent with how \code{\link[gprofiler2:gost]{gprofiler2::gost()}} calculates enrichments.
See \code{\link[MotrpacRatTraining6mo:cluster_pathway_enrichment]{MotrpacRatTraining6mo::cluster_pathway_enrichment()}} for implementation.

For metabolites,
we performed enrichment of KEGG pathways using the hypergeometric method in \code{\link[FELLA:enrich-funs]{FELLA::enrich()}}
with custom backgrounds defined by \link{GENE_UNIVERSES}. See \code{\link[MotrpacRatTraining6mo:run_fella]{MotrpacRatTraining6mo::run_fella()}} for implementation.

Pathway enrichment analysis p-values were adjusted across all results using Independent
Hypothesis Weighting (IHW) with tissue as a covariate.
}
\keyword{datasets}
