% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GROUP_COLORS}
\alias{GROUP_COLORS}
\title{Intervention group colors}
\format{
Named character vector
}
\usage{
GROUP_COLORS
}
\description{
Intervention group colors used in figures. Values are hex codes and names are intervention groups.
}
\keyword{datasets}
