% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_DA}
\alias{IMMUNO_DA}
\alias{IMMUNO_PLASMA_DA}
\alias{IMMUNO_HIPPOC_DA}
\alias{IMMUNO_CORTEX_DA}
\alias{IMMUNO_ADRNL_DA}
\alias{IMMUNO_TESTES_DA}
\alias{IMMUNO_OVARY_DA}
\alias{IMMUNO_SKMGN_DA}
\alias{IMMUNO_BAT_DA}
\alias{IMMUNO_WATSC_DA}
\alias{IMMUNO_COLON_DA}
\alias{IMMUNO_SMLINT_DA}
\alias{IMMUNO_LIVER_DA}
\alias{IMMUNO_SKMVL_DA}
\alias{IMMUNO_HEART_DA}
\alias{IMMUNO_KIDNEY_DA}
\alias{IMMUNO_SPLEEN_DA}
\alias{IMMUNO_LUNG_DA}
\title{Differential analysis of multiplexed immunoassays}
\format{
A data frame with 17 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{dataset}}{character, name of LUMINEX panel, e.g., 'rat-mag27plex'}
\item{\code{sex}}{@eval sex()}
\item{\code{comparison_group}}{@eval comparison_group()}
\item{\code{p_value}}{@eval p_value()}
\item{\code{adj_p_value}}{@eval adj_p_value()}
\item{\code{logFC}}{@eval logFC()}
\item{\code{logFC_se}}{@eval logFC_se()}
\item{\code{covariates}}{@eval covariates()}
\item{\code{comparison_average_intensity}}{@eval comparison_average_intensity()}
\item{\code{reference_average_intensity}}{@eval reference_average_intensity()}
\item{\code{selection_fdr}}{@eval selection_fdr()}
}

An object of class \code{data.frame} with 368 rows and 17 columns.

An object of class \code{data.frame} with 368 rows and 17 columns.

An object of class \code{data.frame} with 368 rows and 17 columns.

An object of class \code{data.frame} with 184 rows and 17 columns.

An object of class \code{data.frame} with 184 rows and 17 columns.

An object of class \code{data.frame} with 424 rows and 17 columns.

An object of class \code{data.frame} with 424 rows and 17 columns.

An object of class \code{data.frame} with 424 rows and 17 columns.

An object of class \code{data.frame} with 280 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 216 rows and 17 columns.

An object of class \code{data.frame} with 208 rows and 17 columns.

An object of class \code{data.frame} with 216 rows and 17 columns.
}
\usage{
IMMUNO_PLASMA_DA

IMMUNO_HIPPOC_DA

IMMUNO_CORTEX_DA

IMMUNO_ADRNL_DA

IMMUNO_TESTES_DA

IMMUNO_OVARY_DA

IMMUNO_SKMGN_DA

IMMUNO_BAT_DA

IMMUNO_WATSC_DA

IMMUNO_COLON_DA

IMMUNO_SMLINT_DA

IMMUNO_LIVER_DA

IMMUNO_SKMVL_DA

IMMUNO_HEART_DA

IMMUNO_KIDNEY_DA

IMMUNO_SPLEEN_DA

IMMUNO_LUNG_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
immunoassay analyte within each sex. One data frame per tissue.
}
\keyword{datasets}
