% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_NORM_DATA_NESTED}
\alias{IMMUNO_NORM_DATA_NESTED}
\title{Processed immunoassay data used for differential analysis}
\format{
A nested list of data frames
}
\source{
\code{gs://mawg-data/pass1b-06/immunoassay/data/release/pass1b-06*_mfi-log2-filt-imputed-na-outliers.txt}
}
\usage{
IMMUNO_NORM_DATA_NESTED
}
\description{
Normalized, imputed, and filtered multiplexed immunoassay data used for differential analysis
}
\details{
IMMUNO sample-level data is in a different format than sample-level data for other assays/omes.
Extract data from a panel and tissue using \code{IMMUNO_SAMPLE_DATA[[panel]][[tissue]]}, where \code{panel}
is one of "ADIPONECTIN", "SERPIN-E", "rat-mag27plex", "rat-metabolic", "rat-myokine", "rat-pituitary", and
\code{tissue} is a tissue abbreviation (see \link{TISSUE_ABBREV}). Samples (vial labels) are in rows, and analytes are
in columns. Column names, barring the first "viallabel" column, correspond to \code{feature_ID}s.

Raw mean fluorescent intensities (MFIs) were log2-transformed, and measurements corresponding to wells with less than 20 beads were removed.
For remaining measurements in each panel and tissue, samples with more than 50\\% missing values (i.e., due to low bead count) were removed;
features with at least two missing values for a single experimental group (e.g., males trained for 2 weeks) were removed
(this affected four colon analytes and one spleen analyte, all in the rat-mag27plex panel). Remaining missing values were
imputed with k-nearest neighbors (k=5 features). Within each panel, tissue, and analyte, we calculated the mean and standard deviation
and removed outlying measurements more than 4 standard deviations away from the mean. This is the version of the data provided by this object
and used for differential analysis.

SERPIN-E and ADIPONECTIN were both originally in a panel called rat-adipokine. They
were split into their own "panels" because this panel was run with two dilutions,
and one was optimal for SERPIN-E while the other was optimal for ADIPONECTIN.
Hence, in some places \code{panel} refers to "rat-adipokine" while in other places,
like here, \code{panel} refers to "SERPIN-E" or "ADIPONECTIN", among others.
}
\keyword{datasets}
