% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_SAMPLE_DATA}
\alias{IMMUNO_SAMPLE_DATA}
\title{Immunoassay normalized sample-level data}
\format{
A data frame with 916 rows and 60 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-targeted/immunoassay-luminex/normalized-data/*merged_mfi-log2-filt-imputed.txt}
}
\usage{
IMMUNO_SAMPLE_DATA
}
\description{
Normalized sample-level multiplexed immunoassay data. Raw mean fluorescent intensities (MFIs)
were log2-transformed, measurements corresponding to wells with less than 20 beads were removed,
and the corresponding missing values were imputed with KNN (k=5).
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
