% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_VIZ_DATA}
\alias{IMMUNO_VIZ_DATA}
\title{Combined immunoassay data used for visualization}
\format{
A data frame with 720 rows and 60 variables:
\describe{
\item{\code{feature}}{unique \code{feature} in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[feature_ID]},
where \code{feature_ID} is \code{new_feature_ID} when applicable. See \link{REPEATED_FEATURES} for details.}
}
}
\source{
\code{gs://mawg-data/pass1b-06/immunoassay/data/release/pass1b-06*_mfi-log2-filt-imputed-na-outliers.txt}
}
\usage{
IMMUNO_VIZ_DATA
}
\description{
Normalized, imputed, and filtered multiplexed immunoassay data used for visualization
}
\details{
Analytes (\code{feature}) are in rows, and participants IDs (PIDs, i.e., unique animal IDs) are in columns.
When an analyte was measured by more than one panel, the panel name is prepended to \code{feature_ID} to
provide unique feature identifiers. See \link{REPEATED_FEATURES} for details.

Data are equivalent to the data provided in \link{IMMUNO_NORM_DATA}. \link{IMMUNO_NORM_DATA} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\keyword{datasets}
