% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_NORM_DATA_FLAT}
\alias{METAB_NORM_DATA_FLAT}
\title{Combined metabolomics data used for visualization}
\format{
A data frame with 59 variables:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link{REPEATED_FEATURES} for details.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{dataset}}{character, metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features.}
}
}
\usage{
METAB_NORM_DATA_FLAT
}
\description{
Combined sample-level metabolomics data used for visualization.
Data are equivalent to the data provided in \link{METAB_SAMPLE_DATA_NESTED}. \link{METAB_SAMPLE_DATA_NESTED} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\details{
TODO
}
\keyword{datasets}
