% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_NORM_DATA_NESTED}
\alias{METAB_NORM_DATA_NESTED}
\title{Processed metabolomics data used for differential analysis}
\format{
A nested list of data frames
}
\usage{
METAB_NORM_DATA_NESTED
}
\description{
Combined sample-level data organized by metabolomics platforms and tissue used for differential analysis
}
\details{
METAB sample-level data is in a different format than sample-level data for other assays/omes.
Extract data from a panel and tissue using \code{METAB_NORM_DATA_NESTED[[platform]][[tissue]]}, where \code{platform}
is one of "metab-t-amines", "metab-t-acoa", "metab-t-nuc", "metab-t-oxylipneg", "metab-t-ka", "metab-t-etamidpos", "metab-t-tca",
"metab-u-lrppos", "metab-u-lrpneg", "metab-u-hilicpos", "metab-u-rppos", "metab-u-rpneg", "metab-u-ionpneg" (see \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes} for details) and
\code{tissue} is a tissue abbreviation (see \link{TISSUE_ABBREV}). Samples (vial labels) are in columns, and metabolites (feature IDs) are
in rows. Feature IDs in the row names in these tables correspond to \code{\link{METAB_FEATURE_ID_MAP}$feature_ID_sample_data}.

TODO: Explain which version of normalized data were used for each data set.
}
\keyword{datasets}
