% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_SAMPLE_DATA}
\alias{METAB_SAMPLE_DATA}
\title{Nested metabolomics data used for differential analysis}
\format{
A tibble with 113 rows and 5 variables:
\describe{
\item{\code{tissue}}{@eval tissue_code()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{sample_data}}{list where first element is a tibble of feature by viallabel normalized sample-level data}
\item{\code{pheno}}{list, phenotypic data important for differential abundance analysis}
\item{\code{feature_metadata}}{list, feature metadata important for differntial abundance analysis}
}
}
\usage{
METAB_SAMPLE_DATA
}
\description{
Combined sample-level data organized by metabolomics platforms and tissue used for differential analysis
}
\details{
TODO
}
\keyword{datasets}
