% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{METHYL_DA}
\alias{METHYL_DA}
\title{Differential analysis of RRBS data}
\format{
TODO
}
\description{
TODO
}
\details{
TODO
Full METHYL differential analysis results are only available via download from the Cloud. See TODO.

For training-regulated METHYL features at 5\% FDR, see \link{TRAINING_REGULATED_FEATURES}.
}
