% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METHYL_META}
\alias{METHYL_META}
\title{RRBS metadata and QC}
\format{
A data frame with 416 rows and 75 variables:
\describe{
\item{\code{viallabel}}{@eval viallabel()}
\item{\code{vial_label}}{double, sample identifier}
\item{\code{2D_barcode}}{double, sample barcode}
\item{\code{Species}}{character, species}
\item{\code{BID}}{integer, biospecimen ID}
\item{\code{PID}}{double, participant ID, one per animal}
\item{\code{Tissue}}{character, tissue description}
\item{\code{Sample_category}}{character, study sample ("study") or reference standard ("ref)}
\item{\code{GET_site}}{character, which Genomics, Epigenomics, and Transcriptomics (GET)
site performed the assay, "Stanford" or "MSSM" (Icahn School of Medicine at Mount Sinai)}
\item{\code{DNA_extr_plate_ID}}{integer, DNA extraction plate ID}
\item{\code{DNA_extr_date}}{character, DNA extraction date}
\item{\code{DNA_extr_protocol}}{character, DNA extraction protocol}
\item{\code{DNA_extr_robot}}{character, robot used for DNA extraction}
\item{\code{DNA_conc}}{double, DNA concentration (ng/uL)}
\item{\code{A280/260}}{double, 280/260 ratio}
\item{\code{A260/230}}{double, 260/230 ratio}
\item{\code{Lib_prep_date}}{character, library preparation date}
\item{\code{Lib_DNA_mass}}{double, DNA mass used for library prep (ng)}
\item{\code{Lib_DNA_vol}}{double, volume of the library (uL)}
\item{\code{lambda_DNA_mass}}{double, spiked-in Lambda DNA mass (ng)}
\item{\code{Lib_robot}}{character, robot used for library prep}
\item{\code{Lib_kit_vendor}}{character, library prep vendor}
\item{\code{Lib_kit_type}}{character, library prep kit}
\item{\code{Lib_kit_ID}}{character, library kit ID}
\item{\code{Lib_batch_ID}}{character, library prep batch ID}
\item{\code{Lib_index_1}}{character, i7 index}
\item{\code{Lib_index_2}}{logical, i5 index}
\item{\code{Lib_adapter_1}}{character, Trueseq i7 index with 16 bp index}
\item{\code{Lib_adapter_2}}{character, include customized Metseq primer}
\item{\code{Lib_UMI_cycle_num}}{integer, number of bases of UMI}
\item{\code{Lib_adapter_size}}{integer, the total size of the two adapters}
\item{\code{Lib_conc}}{double, DNA concentration for the library (ng/uL)}
\item{\code{Lib_frag_size}}{integer, average library fragment size}
\item{\code{Lib_molarity}}{double, library molarity (nM)}
\item{\code{Seq_platform}}{character, sequencing platform}
\item{\code{Seq_date}}{integer, sequencing date}
\item{\code{Seq_machine_ID}}{character, serial number of the sequencing machine}
\item{\code{Seq_flowcell_ID}}{character, flow cell ID}
\item{\code{Seq_flowcell_run}}{integer, flow cell run}
\item{\code{Seq_flowcell_lane}}{character, flow cell lane}
\item{\code{Seq_flowcell_type}}{character, flow cell type, e.g., "S4"}
\item{\code{Seq_length}}{integer, read length}
\item{\code{Seq_end_type}}{integer, 1=single-end, 2=paired-end}
\item{\code{reads_raw}}{integer, number of raw read pairs}
\item{\code{pct_adapter_detected}}{double, percent of reads with adapter detected}
\item{\code{pct_trimmed}}{double, percent of trimmed reads from the initial trimming}
\item{\code{pct_no_MSPI}}{double, percent of reads with no MSPI present among the trimmed reads}
\item{\code{pct_trimmed_bases}}{double, percent of bases that were trimmed}
\item{\code{pct_removed}}{double, percent of reads that were removed due to adapter trimming or no presence of MSPI}
\item{\code{reads}}{integer, number of read pairs in the trimmed FASTQ files}
\item{\code{pct_GC}}{double, percent GC content in the trimmed FASTQ files}
\item{\code{pct_dup_sequence}}{double, percent of duplicated sequences in trimmed FASTQ files}
\item{\code{pct_phix}}{double, percent of phix reads in trimmed FASTQ files}
\item{\code{pct_chrX}}{double, percent of reads mapped to chromosome X}
\item{\code{pct_chrY}}{double, percent of reads mapped to chromosome Y}
\item{\code{pct_chrM}}{double, percent of reads mapped to the mitochondrial genome}
\item{\code{pct_chrAuto}}{double, percent of reads mapped to autosomal chromosomes}
\item{\code{pct_contig}}{double, percent of reads mapped to contigs}
\item{\code{pct_Uniq}}{double, percent of uniquely mapped reads}
\item{\code{pct_Unaligned}}{double, percent of unaligned reads}
\item{\code{pct_Ambi}}{double, percent of ambiguously mapped reads}
\item{\code{pct_OT}}{double, percent of mapped reads aligned to the original top stand}
\item{\code{pct_OB}}{double, percent of mapped reads aligned to the original bottom stand}
\item{\code{pct_CTOT}}{double, percent of mapped reads aligned to the complementary to original top strand}
\item{\code{pct_CTOB}}{double, percent of mapped reads aligned to the complementary to original bottom strand}
\item{\code{pct_umi_dup}}{double, PCR duplication rate assessed using UMIs (Unique Molecular Identifiers)}
\item{\code{pct_CpG}}{double, global CpG methylation level based on the deduplicated data}
\item{\code{pct_CHG}}{double, global CHG methylation level  based on the deduplicated data}
\item{\code{pct_CHH}}{double, global CHH methylation level  based on the deduplicated data}
\item{\code{lambda_pct_Uniq}}{double, percent of uniquely mapped reads to lambda}
\item{\code{lambda_pct_Ambi}}{double, percent of ambiguously mapped reads to lambda}
\item{\code{lambda_pct_umi_dup}}{double, PCR duplication rate assessed using UMIs (Unique Molecular Identifiers)}
\item{\code{lambda_pct_CpG}}{double, global CpG methylation level based on the deduplicated data}
\item{\code{lambda_pct_CHG}}{double, global CHG methylation level  based on the deduplicated data}
\item{\code{lambda_pct_CHH}}{double, global CHH methylation level  based on the deduplicated data}
\item{\code{Seq_batch}}{character, unique identifier for sequencing batch}
}
}
\source{
\url{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/results/epigenomics/qa-qc/motrpac_pass1b-06_epigen-rrbs_qa-qc-metrics.csv}
}
\usage{
METHYL_META
}
\description{
RRBS experimental and quantification QC metrics for DNA methylation (METHYL) data
}
\keyword{datasets}
