% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METHYL_SAMPLE_DATA}
\alias{METHYL_SAMPLE_DATA}
\title{RRBS normalized sample-level data for training-regulated features}
\format{
A data frame with 1530 rows and 51 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/epigenomics/epigen-rrbs/normalized-data/*normalized-log-M-window.txt*}
}
\usage{
METHYL_SAMPLE_DATA
}
\description{
Normalized RRBS sample-level data for training-regulated features only.
See the data release on \url{https://motrpac-data.org/} for sample-level data for all features.
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
