% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OUTLIERS}
\alias{OUTLIERS}
\title{Sample outliers}
\format{
A data frame with 27 rows and 4 variables:
\describe{
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{viallabel}}{double, sample identifier}
\item{\code{reason}}{character, reason(s) the sample was called an outlier}
}
}
\usage{
OUTLIERS
}
\description{
Outliers excluded during differential analysis
}
\details{
If the sample was an outlier in principal component (PC) space, \code{reason}
lists the PC(s) in which it was an outlier. See ome-specific details of outlier calling
in the supplementary methods of the manuscript.
}
\keyword{datasets}
