% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PATHWAY_PARENTS}
\alias{PATHWAY_PARENTS}
\title{KEGG and Reactome parent pathways}
\format{
List of length 21697
}
\usage{
PATHWAY_PARENTS
}
\description{
List where names are KEGG or Reactome terms IDs and values are
the parent pathway(s)
}
\details{
Pathway hierarchies for KEGG were retrieved with \code{\link[KEGGREST:keggGet]{KEGGREST::keggGet()}}, e.g. \code{keggGet(kegg_query)[[1]]$CLASS}.
Reactome pathway parents were extracted from \url{https://reactome.org/download/current/ReactomePathwaysRelation.txt}.
}
\keyword{datasets}
