% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHENO}
\alias{PHENO}
\title{Phenotypic data}
\format{
A data frame with 5955 rows and 510 variables:
\describe{
\item{\code{pid}}{integer, randomly generated 8-digit numeric identifier used in linkage to phenotypic data.
One PID per animal, i.e., all samples from the same animal have the same PID.}
\item{\code{bid}}{integer, 5-digit numeric identifier for all samples collected for a sample collection period.
All samples collected during that period will have the same BID.}
\item{\code{labelid}}{double, unique 11 digit identifier for specimen label ID, originating at the collection site,
that provides a link to specimen processing and used for shipments to the biorepository (same as
\code{viallabel} only in instances where aliquots were not further processed at the biorepository)}
\item{\code{viallabel}}{character, unique 11 digit numeric identifier of a sample vial made up of \code{bid},
\code{key.sacrificetime}, \code{specimen.processing.sampletypedescription}, and aliquot number}
\item{\code{key.protocol}}{character, protocol in which the procedures for sample collection were outlined}
\item{\code{key.agegroup}}{character, age group for the animal}
\item{\code{key.d_arrive}}{character, date animal arrived at site (same as \code{registration.d_arrive})}
\item{\code{key.d_sacrifice}}{character, date animal was sacrificed (same as \code{specimen.collection.d_visit})}
\item{\code{key.intervention}}{character, exercise group defined by those who complete exercise training}
\item{\code{key.sacrificetime}}{character, weeks since the beginning of the intervention}
\item{\code{key.anirandgroup}}{character, description of the randomization group}
\item{\code{key.siteid}}{character, site where the samples were collected}
\item{\code{registration.d_visit}}{character, date registration was completed}
\item{\code{registration.days_visit}}{integer, day registration was completed relative to \code{registration.d_arrive} (count)}
\item{\code{registration.staffid}}{integer, registration staff ID}
\item{\code{registration.siteid}}{character, registration site}
\item{\code{registration.ratid}}{character, site rat ID}
\item{\code{registration.d_arrive}}{character, date the rat arrived at the site}
\item{\code{registration.days_arrive}}{integer, day of arrival relative to \code{registration.d_arrive} (count)}
\item{\code{registration.batchnumber}}{integer, registration batch number}
\item{\code{registration.d_reverselight}}{character, date the rat was housed in reverse light}
\item{\code{registration.days_reverselight}}{integer, day the rat was housed in reverse light relative to \code{registration.d_arrive} (count)}
\item{\code{registration.d_birth}}{character, rat date of birth (+/- two weeks)}
\item{\code{registration.days_birth}}{integer, rat day of birth relative to \code{registration.d_arrive} (count)}
\item{\code{registration.sex}}{character, sex, one of "male", "female"}
\item{\code{registration.weight}}{double, rat's body weight}
\item{\code{registration.cagenumber}}{integer, cage number in which the rat was housed}
\item{\code{registration.comments}}{character, registration comments}
\item{\code{familiarization.d_visit}}{character, date familiarization was completed}
\item{\code{familiarization.days_visit}}{integer, day familiarization was completed relative to \code{registration.d_arrive} (count)}
\item{\code{familiarization.staffid}}{integer, familiarization staff ID}
\item{\code{familiarization.siteid}}{character, familiarization site}
\item{\code{familiarization.d_treadmillbegin}}{character, date the rat began treadmill familiarization}
\item{\code{familiarization.days_treadmillbegin}}{integer, day the rat began treadmill familiarization relative to \code{registration.d_arrive} (count)}
\item{\code{familiarization.d_treadmillcomplete}}{character, date the rat completed treadmill familiarization}
\item{\code{familiarization.days_treadmillcomplete}}{integer, day the rat completed treadmill familiarization relative to \code{registration.d_arrive} (count)}
\item{\code{familiarization.activity_score}}{double, rat's activity score}
\item{\code{familiarization.compliant}}{character, was this rat compliant?}
\item{\code{familiarization.weight}}{double, rat's body weight}
\item{\code{training.d_visit}}{character, date intervention started}
\item{\code{training.days_visit}}{integer, day intervention started relative to \code{registration.d_arrive}}
\item{\code{training.staffid}}{integer, training staff ID}
\item{\code{training.siteid}}{character, training site}
\item{\code{training.day1date}}{character, date of given training day}
\item{\code{training.day1_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day1time}}{character, time of given training day}
\item{\code{training.day1_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day1_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day1_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day1_weight}}{double, weight on given training day}
\item{\code{training.day1_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day1_score}}{integer, score on given training day}
\item{\code{training.day1_comments}}{character, comments on given training day}
\item{\code{training.day2date}}{character, date of given training day}
\item{\code{training.day2_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day2time}}{character, time of given training day}
\item{\code{training.day2_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day2_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day2_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day2_score}}{integer, score on given training day}
\item{\code{training.day2_comments}}{character, comments on given training day}
\item{\code{training.day3date}}{character, date of given training day}
\item{\code{training.day3_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day3time}}{character, time of given training day}
\item{\code{training.day3_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day3_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day3_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day3_score}}{integer, score on given training day}
\item{\code{training.day3_comments}}{character, comments on given training day}
\item{\code{training.day4date}}{character, date of given training day}
\item{\code{training.day4_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day4time}}{character, time of given training day}
\item{\code{training.day4_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day4_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day4_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day4_score}}{integer, score on given training day}
\item{\code{training.day4_comments}}{character, comments on given training day}
\item{\code{training.day5date}}{character, date of given training day}
\item{\code{training.day5_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day5time}}{character, time of given training day}
\item{\code{training.day5_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day5_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day5_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day5_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day5_score}}{integer, score on given training day}
\item{\code{training.day5_comments}}{character, comments on given training day}
\item{\code{training.day6date}}{character, date of given training day}
\item{\code{training.day6_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day6time}}{character, time of given training day}
\item{\code{training.day6_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day6_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day6_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day6_weight}}{double, weight on given training day}
\item{\code{training.day6_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day6_score}}{integer, score on given training day}
\item{\code{training.day6_comments}}{character, comments on given training day}
\item{\code{training.day7date}}{character, date of given training day}
\item{\code{training.day7_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day7time}}{character, time of given training day}
\item{\code{training.day7_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day7_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day7_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day7_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day7_score}}{integer, score on given training day}
\item{\code{training.day7_comments}}{character, comments on given training day}
\item{\code{training.day8date}}{character, date of given training day}
\item{\code{training.day8_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day8time}}{character, time of given training day}
\item{\code{training.day8_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day8_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day8_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day8_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day8_score}}{integer, score on given training day}
\item{\code{training.day8_comments}}{character, comments on given training day}
\item{\code{training.day9date}}{character, date of given training day}
\item{\code{training.day9_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day9time}}{character, time of given training day}
\item{\code{training.day9_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day9_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day9_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day9_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day9_score}}{integer, score on given training day}
\item{\code{training.day9_comments}}{character, comments on given training day}
\item{\code{training.day10date}}{character, date of given training day}
\item{\code{training.day10_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day10time}}{character, time of given training day}
\item{\code{training.day10_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day10_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day10_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day10_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day10_score}}{integer, score on given training day}
\item{\code{training.day10_comments}}{character, comments on given training day}
\item{\code{training.day11date}}{character, date of given training day}
\item{\code{training.day11_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day11time}}{character, time of given training day}
\item{\code{training.day11_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day11_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day11_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day11_weight}}{double, weight on given training day}
\item{\code{training.day11_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day11_score}}{integer, score on given training day}
\item{\code{training.day11_comments}}{character, comments on given training day}
\item{\code{training.day12date}}{character, date of given training day}
\item{\code{training.day12_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day12time}}{character, time of given training day}
\item{\code{training.day12_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day12_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day12_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day12_score}}{integer, score on given training day}
\item{\code{training.day12_comments}}{character, comments on given training day}
\item{\code{training.day13date}}{character, date of given training day}
\item{\code{training.day13_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day13time}}{character, time of given training day}
\item{\code{training.day13_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day13_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day13_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day13_score}}{integer, score on given training day}
\item{\code{training.day13_comments}}{character, comments on given training day}
\item{\code{training.day14date}}{character, date of given training day}
\item{\code{training.day14_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day14time}}{character, time of given training day}
\item{\code{training.day14_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day14_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day14_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day14_score}}{integer, score on given training day}
\item{\code{training.day14_comments}}{character, comments on given training day}
\item{\code{training.day15date}}{character, date of given training day}
\item{\code{training.day15_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day15time}}{character, time of given training day}
\item{\code{training.day15_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day15_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day15_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day15_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day15_score}}{integer, score on given training day}
\item{\code{training.day15_comments}}{character, comments on given training day}
\item{\code{training.day16date}}{character, date of given training day}
\item{\code{training.day16_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day16time}}{character, time of given training day}
\item{\code{training.day16_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day16_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day16_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day16_weight}}{double, weight on given training day}
\item{\code{training.day16_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day16_score}}{integer, score on given training day}
\item{\code{training.day16_comments}}{character, comments on given training day}
\item{\code{training.day17date}}{character, date of given training day}
\item{\code{training.day17_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day17time}}{character, time of given training day}
\item{\code{training.day17_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day17_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day17_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day17_score}}{integer, score on given training day}
\item{\code{training.day17_comments}}{character, comments on given training day}
\item{\code{training.day18date}}{character, date of given training day}
\item{\code{training.day18_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day18time}}{character, time of given training day}
\item{\code{training.day18_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day18_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day18_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day18_score}}{integer, score on given training day}
\item{\code{training.day18_comments}}{character, comments on given training day}
\item{\code{training.day19date}}{character, date of given training day}
\item{\code{training.day19_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day19time}}{character, time of given training day}
\item{\code{training.day19_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day19_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day19_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day19_score}}{integer, score on given training day}
\item{\code{training.day19_comments}}{character, comments on given training day}
\item{\code{training.day20date}}{character, date of given training day}
\item{\code{training.day20_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day20time}}{character, time of given training day}
\item{\code{training.day20_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day20_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day20_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day20_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day20_score}}{integer, score on given training day}
\item{\code{training.day20_comments}}{character, comments on given training day}
\item{\code{training.day21date}}{character, date of given training day}
\item{\code{training.day21_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day21time}}{character, time of given training day}
\item{\code{training.day21_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day21_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day21_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day21_weight}}{double, weight on given training day}
\item{\code{training.day21_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day21_score}}{integer, score on given training day}
\item{\code{training.day21_comments}}{character, comments on given training day}
\item{\code{training.day22date}}{character, date of given training day}
\item{\code{training.day22_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day22time}}{character, time of given training day}
\item{\code{training.day22_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day22_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day22_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day22_score}}{integer, score on given training day}
\item{\code{training.day22_comments}}{character, comments on given training day}
\item{\code{training.day23date}}{character, date of given training day}
\item{\code{training.day23_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day23time}}{character, time of given training day}
\item{\code{training.day23_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day23_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day23_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day23_score}}{integer, score on given training day}
\item{\code{training.day23_comments}}{character, comments on given training day}
\item{\code{training.day24date}}{character, date of given training day}
\item{\code{training.day24_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day24time}}{character, time of given training day}
\item{\code{training.day24_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day24_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day24_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day24_score}}{integer, score on given training day}
\item{\code{training.day24_comments}}{character, comments on given training day}
\item{\code{training.day25date}}{character, date of given training day}
\item{\code{training.day25_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day25time}}{character, time of given training day}
\item{\code{training.day25_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day25_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day25_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day25_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day25_score}}{integer, score on given training day}
\item{\code{training.day25_comments}}{character, comments on given training day}
\item{\code{training.day26date}}{character, date of given training day}
\item{\code{training.day26_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day26time}}{character, time of given training day}
\item{\code{training.day26_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day26_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day26_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day26_weight}}{double, weight on given training day}
\item{\code{training.day26_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day26_score}}{integer, score on given training day}
\item{\code{training.day26_comments}}{character, comments on given training day}
\item{\code{training.day27date}}{character, date of given training day}
\item{\code{training.day27_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day27time}}{character, time of given training day}
\item{\code{training.day27_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day27_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day27_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day27_score}}{integer, score on given training day}
\item{\code{training.day27_comments}}{character, comments on given training day}
\item{\code{training.day28date}}{character, date of given training day}
\item{\code{training.day28_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day28time}}{character, time of given training day}
\item{\code{training.day28_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day28_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day28_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day28_score}}{integer, score on given training day}
\item{\code{training.day28_comments}}{character, comments on given training day}
\item{\code{training.day29date}}{character, date of given training day}
\item{\code{training.day29_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day29time}}{character, time of given training day}
\item{\code{training.day29_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day29_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day29_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day29_score}}{integer, score on given training day}
\item{\code{training.day29_comments}}{character, comments on given training day}
\item{\code{training.day30date}}{character, date of given training day}
\item{\code{training.day30_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day30time}}{character, time of given training day}
\item{\code{training.day30_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day30_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day30_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day30_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day30_score}}{integer, score on given training day}
\item{\code{training.day30_comments}}{character, comments on given training day}
\item{\code{training.day31date}}{character, date of given training day}
\item{\code{training.day31_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day31time}}{character, time of given training day}
\item{\code{training.day31_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day31_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day31_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day31_weight}}{double, weight on given training day}
\item{\code{training.day31_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day31_score}}{integer, score on given training day}
\item{\code{training.day31_comments}}{character, comments on given training day}
\item{\code{training.day32date}}{character, date of given training day}
\item{\code{training.day32_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day32time}}{character, time of given training day}
\item{\code{training.day32_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day32_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day32_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day32_score}}{integer, score on given training day}
\item{\code{training.day32_comments}}{character, comments on given training day}
\item{\code{training.day33date}}{character, date of given training day}
\item{\code{training.day33_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day33time}}{character, time of given training day}
\item{\code{training.day33_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day33_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day33_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day33_score}}{integer, score on given training day}
\item{\code{training.day33_comments}}{character, comments on given training day}
\item{\code{training.day34date}}{character, date of given training day}
\item{\code{training.day34_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day34time}}{character, time of given training day}
\item{\code{training.day34_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day34_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day34_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day34_score}}{integer, score on given training day}
\item{\code{training.day34_comments}}{character, comments on given training day}
\item{\code{training.day35date}}{character, date of given training day}
\item{\code{training.day35_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day35time}}{character, time of given training day}
\item{\code{training.day35_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day35_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day35_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day35_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day35_score}}{integer, score on given training day}
\item{\code{training.day35_comments}}{character, comments on given training day}
\item{\code{training.day36date}}{character, date of given training day}
\item{\code{training.day36_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day36time}}{character, time of given training day}
\item{\code{training.day36_weight}}{double, weight on given training day}
\item{\code{training.day36_comments}}{character, comments on given training day}
\item{\code{training.day37date}}{character, date of given training day}
\item{\code{training.day37_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day37time}}{character, time of given training day}
\item{\code{training.day37_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day37_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day37_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day37_score}}{integer, score on given training day}
\item{\code{training.day37_comments}}{character, comments on given training day}
\item{\code{training.day38date}}{character, date of given training day}
\item{\code{training.day38_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day38time}}{character, time of given training day}
\item{\code{training.day38_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day38_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day38_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day38_score}}{integer, score on given training day}
\item{\code{training.day38_comments}}{character, comments on given training day}
\item{\code{training.day39date}}{character, date of given training day}
\item{\code{training.day39_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day39time}}{character, time of given training day}
\item{\code{training.day39_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day39_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day39_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day39_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day39_score}}{integer, score on given training day}
\item{\code{training.day39_comments}}{character, comments on given training day}
\item{\code{training.day40date}}{character, date of given training day}
\item{\code{training.day40_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day40time}}{character, time of given training day}
\item{\code{training.day40_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day40_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day40_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day40_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day40_score}}{integer, score on given training day}
\item{\code{training.day40_comments}}{character, comments on given training day}
\item{\code{vo2.max.test.d_visit_1}}{character, date first VO2 max test was completed}
\item{\code{vo2.max.test.d_visit_2}}{character, date second VO2 max test was completed}
\item{\code{vo2.max.test.days_visit_1}}{integer, day first VO2 max test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.days_visit_2}}{integer, day second VO2 max test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.staffid_1}}{integer, first VO2 max test staff ID}
\item{\code{vo2.max.test.staffid_2}}{integer, second VO2 max test staff ID}
\item{\code{vo2.max.test.siteid_1}}{character, first VO2 max test site}
\item{\code{vo2.max.test.siteid_2}}{character, second VO2 max test site}
\item{\code{vo2.max.test.visitcode_1}}{character, visit code for first VO2 max test}
\item{\code{vo2.max.test.visitcode_2}}{character, visit code for second VO2 max test}
\item{\code{vo2.max.test.formname_1}}{character, form name for first VO2 max test}
\item{\code{vo2.max.test.formname_2}}{character, form name for second VO2 max test}
\item{\code{vo2.max.test.versionnbr_1}}{integer, form version number for first VO2 max test}
\item{\code{vo2.max.test.versionnbr_2}}{integer, form version number for second VO2 max test}
\item{\code{vo2.max.test.d_vo2_1}}{character, date of first VO2 max test}
\item{\code{vo2.max.test.d_vo2_2}}{character, date of second VO2 max test}
\item{\code{vo2.max.test.days_vo2_1}}{integer, date of first VO2 max test relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.days_vo2_2}}{integer, date of second VO2 max test relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.t_vo2_1}}{character, start time of first VO2 max test?}
\item{\code{vo2.max.test.t_vo2_2}}{character, start time of second VO2 max test?}
\item{\code{vo2.max.test.blactate_begin_1}}{double, blood lactate at the beginning of exercise for first VO2 max test}
\item{\code{vo2.max.test.blactate_begin_2}}{double, blood lactate at the beginning of exercise for second VO2 max test}
\item{\code{vo2.max.test.vo2_max_1}}{double, VO2 at max for first VO2 max test}
\item{\code{vo2.max.test.vo2_max_2}}{double, VO2 at max for second VO2 max test}
\item{\code{vo2.max.test.vco2_max_1}}{double, VCO2 at max for first VO2 max test}
\item{\code{vo2.max.test.vco2_max_2}}{double, VCO2 at max for second VO2 max test}
\item{\code{vo2.max.test.rer_max_1}}{double, RER at max for first VO2 max test}
\item{\code{vo2.max.test.rer_max_2}}{double, RER at max for second VO2 max test}
\item{\code{vo2.max.test.speed_max_1}}{double, max speed of running for first VO2 max test}
\item{\code{vo2.max.test.speed_max_2}}{double, max speed of running for second VO2 max test}
\item{\code{vo2.max.test.t_complete_1}}{character, time of completion for first VO2 max test}
\item{\code{vo2.max.test.t_complete_2}}{character, time of completion for second VO2 max test}
\item{\code{vo2.max.test.blactate_end_1}}{double, blood lactate at end of exercise for first VO2 max test}
\item{\code{vo2.max.test.blactate_end_2}}{double, blood lactate at end of exercise for second VO2 max test}
\item{\code{vo2.max.test.vo2_comments_1}}{character, comments about first VO2 max test}
\item{\code{vo2.max.test.vo2_comments_2}}{character, comments about second VO2 max test}
\item{\code{nmr.testing.d_visit_1}}{character, date first NMR test was completed}
\item{\code{nmr.testing.d_visit_2}}{character, date second NMR test was completed}
\item{\code{nmr.testing.days_visit_1}}{integer, day first NMR test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.days_visit_2}}{integer, day second NMR test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.staffid_1}}{integer, first NMR test staff ID}
\item{\code{nmr.testing.staffid_2}}{integer, second NMR test staff ID}
\item{\code{nmr.testing.siteid_1}}{character, first NMR test site}
\item{\code{nmr.testing.siteid_2}}{character, second NMR test site}
\item{\code{nmr.testing.visitcode_1}}{character, visit code for first NMR test}
\item{\code{nmr.testing.visitcode_2}}{character, visit code for second NMR test}
\item{\code{nmr.testing.formname_1}}{character, form name for first NMR test}
\item{\code{nmr.testing.formname_2}}{character, form name for second NMR test}
\item{\code{nmr.testing.versionnbr_1}}{integer, form version number for first NMR test}
\item{\code{nmr.testing.versionnbr_2}}{integer, form version number for second NMR test}
\item{\code{nmr.testing.d_nmr_1}}{character, date data was collected for the first NMR test}
\item{\code{nmr.testing.d_nmr_2}}{character, date data was collected for the second NMR test}
\item{\code{nmr.testing.days_nmr_1}}{integer, day data was collected for the first NMR test relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.days_nmr_2}}{integer, day data was collected for the second NMR test relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.nmr_weight_1}}{double, rat's body weight for first NMR test}
\item{\code{nmr.testing.nmr_weight_2}}{double, rat's body weight for second NMR test}
\item{\code{nmr.testing.nmr_fat_1}}{double, rat's percent body fat for first NMR test}
\item{\code{nmr.testing.nmr_fat_2}}{double, rat's percent body fat for second NMR test}
\item{\code{nmr.testing.nmr_lean_1}}{double, rat's percent lean for first NMR test}
\item{\code{nmr.testing.nmr_lean_2}}{double, rat's percent lean for second NMR test}
\item{\code{nmr.testing.nmr_fluid_1}}{double, rat's percent fluid for first NMR test}
\item{\code{nmr.testing.nmr_fluid_2}}{double, rat's percent fluid for second NMR test}
\item{\code{nmr.testing.nmr_comments_2}}{character, comments about second NMR test}
\item{\code{terminal.weight.bw}}{double, terminal body weight (g)}
\item{\code{terminal.weight.lg}}{integer, terminal lateral gastrocnemius weight (mg)}
\item{\code{terminal.weight.mg}}{integer, terminal medial gastrocnemius weight (mg)}
\item{\code{terminal.weight.pl}}{integer, terminal plantaris weight (mg)}
\item{\code{terminal.weight.sol}}{integer, terminal soleus weight (mg)}
\item{\code{specimen.collection.d_visit}}{character, date of specimen collection}
\item{\code{specimen.collection.days_visit}}{integer, day of specimen collection relative to \code{registration.d_arrive} (count)}
\item{\code{specimen.collection.staffid}}{integer, specimen collection staff ID}
\item{\code{specimen.collection.siteid}}{character, specimen collection site}
\item{\code{specimen.collection.visitcode}}{character, specimen collection site visit code}
\item{\code{specimen.collection.anesthesiaid}}{integer, anesthesia administrator ID}
\item{\code{specimen.collection.t_anesthesia}}{character, anesthesia administration time}
\item{\code{specimen.collection.bloodtype}}{character, blood type}
\item{\code{specimen.collection.bloodtube}}{character, blood tube}
\item{\code{specimen.collection.bloodcomplete}}{character, blood collection complete}
\item{\code{specimen.collection.bloodtechid}}{character, blood collection technician ID}
\item{\code{specimen.collection.t_bloodstart}}{character, blood collection start time}
\item{\code{specimen.collection.t_bloodstop}}{character, blood collection stop time}
\item{\code{specimen.collection.t_edtafill}}{character, EDTA fill time}
\item{\code{specimen.collection.bloodcomments}}{character, comments about blood collection}
\item{\code{specimen.collection.uterustype}}{character, uterus type}
\item{\code{specimen.collection.uteruscomplete}}{character, uterus complete}
\item{\code{specimen.collection.t_death}}{character, time of death}
\item{\code{specimen.collection.deathtype}}{character, type of death}
\item{\code{specimen.processing.versionnbr}}{integer, specimen processing form version number}
\item{\code{specimen.processing.formname}}{character, specimen processing form name}
\item{\code{specimen.processing.visitcode}}{character, specimen processing visit code}
\item{\code{specimen.processing.siteid}}{character, specimen processing site}
\item{\code{specimen.processing.sampletypedescription}}{character, specimen tissue type}
\item{\code{specimen.processing.samplenumber}}{integer, specimen sample number}
\item{\code{specimen.processing.timepoint}}{integer, specimen time point}
\item{\code{specimen.processing.aliquotdescription}}{character, specimen aliquot tube}
\item{\code{specimen.processing.volume}}{character, specimen aliquot volume}
\item{\code{specimen.processing.hemolyzed}}{character, specimen hemolyzed?}
\item{\code{specimen.processing.comments}}{character, specimen processing comments}
\item{\code{specimen.processing.t_collection}}{character, specimen collection time}
\item{\code{specimen.processing.t_edtaspin}}{character, EDTA spin time}
\item{\code{specimen.processing.t_freeze}}{character, specimen freeze time}
\item{\code{specimen.processing.techid}}{character, specimen processing tech ID}
\item{\code{calculated.variables.pct_body_fat_change}}{double, percent body fat change after training:
\code{nmr.testing.nmr_fat_2 - nmr.testing.nmr_fat_1}}
\item{\code{calculated.variables.pct_body_lean_change}}{double, percent body lean change after training:
\code{nmr.testing.nmr_lean_2 - nmr.testing.nmr_lean_1}}
\item{\code{calculated.variables.pct_body_fluid_change}}{double, percent body fluid change after training:
\code{nmr.testing.nmr_fluid_2 - nmr.testing.nmr_fluid_1}}
\item{\code{calculated.variables.lactate_change_dueto_train}}{double, lactate change due to training:
\code{vo2.max.test.blactate_end_2 - vo2.max.test.blactate_begin_1}}
\item{\code{calculated.variables.vo2_max_change}}{double, change in VO2 max after training:
\code{vo2.max.test.vo2_max_2 - vo2.max.test.vo2_max_1}}
\item{\code{calculated.variables.coll_time_train}}{integer, sample collection time after training (seconds):
specimen collection date and time - last time trained}
\item{\code{calculated.variables.deathtime_after_train}}{integer, time of death after training (seconds):
death date and time - last time trained}
\item{\code{calculated.variables.frozetime_after_train}}{integer, time of sample freezing after training (seconds):
specimen collection date and freeze time - last time trained}
\item{\code{biclabeldata.barcode}}{double, scannable tube identifier used to map samples to vial labels}
\item{\code{biclabeldata.shiptositeid}}{character, Chemical Analysis Site (CAS), i.e. the omics site)}
\item{\code{biclabeldata.received_cas}}{character, was the sample recieved by the Chemical Analysis Site (CAS)?}
\item{\code{biclabeldata.protocol}}{character, study protocol ("phase 1b")}
\item{\code{study_group_timepoint}}{character, a BIC-created field that combines the intervention and sacrifice time codes}
\item{\code{sacrificetime}}{custom column: character, sacrifice time point in weeks, one of "1w", "2w", "4w", "8w"}
\item{\code{intervention}}{custom column: character, intervention, one of "training", "control"}
\item{\code{group}}{custom column: character, intervention group, one of "1w", "2w", "4w", "8w", "control"}
\item{\code{sex}}{custom column: character, one of 'male' or 'female'}
\item{\code{time_to_freeze}}{custom column: integer, sample time from death to freeze, \code{calculated.variables.frozetime_after_train - calculated.variables.deathtime_after_train}}
\item{\code{tissue_code_no}}{custom column: character, BIC tissue code. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{tissue_description}}{custom column: character, original tissue description (from column \code{specimen.processing.sampletypedescription})}
\item{\code{tissue}}{custom column: character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
}
}
\usage{
PHENO
}
\description{
Phenotypic data for samples from the MoTrPAC endurance exercise
training study in 6-month-old rats. One row per sample (\code{viallabel}).
}
\details{
Updated to v3 on 12/27/22. Relative to the original phenotypic data available
through the \href{https://motrpac-data.org/}{MoTrPAC Data Portal},
all-NA columns are removed, and column names are reverted to a previous version
for backwards compatibility. Several columns are also added for simplicity.
}
\keyword{datasets}
