% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHOSPHO_NORM_DATA}
\alias{PHOSPHO_NORM_DATA}
\alias{PHOSPHO_CORTEX_NORM_DATA}
\alias{PHOSPHO_SKMGN_NORM_DATA}
\alias{PHOSPHO_HEART_NORM_DATA}
\alias{PHOSPHO_KIDNEY_NORM_DATA}
\alias{PHOSPHO_LUNG_NORM_DATA}
\alias{PHOSPHO_LIVER_NORM_DATA}
\alias{PHOSPHO_WATSC_NORM_DATA}
\title{Normalized protein phosphorylation data}
\format{
A data frame with phosphosites in rows (\code{feature_ID}) and samples in columns (\code{viallabel})

An object of class \code{data.frame} with 29651 rows and 58 columns.

An object of class \code{data.frame} with 40208 rows and 61 columns.

An object of class \code{data.frame} with 30144 rows and 61 columns.

An object of class \code{data.frame} with 47661 rows and 61 columns.

An object of class \code{data.frame} with 43567 rows and 60 columns.

An object of class \code{data.frame} with 30304 rows and 61 columns.
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-untargeted/prot-ph/normalized-data/*med-mad-normalized-logratio.txt}
}
\usage{
PHOSPHO_CORTEX_NORM_DATA

PHOSPHO_SKMGN_NORM_DATA

PHOSPHO_HEART_NORM_DATA

PHOSPHO_KIDNEY_NORM_DATA

PHOSPHO_LUNG_NORM_DATA

PHOSPHO_LIVER_NORM_DATA

PHOSPHO_WATSC_NORM_DATA
}
\description{
Median-MAD normalized protein phosphorylation (PHOSPHO) data used for visualization and differential analysis
}
\details{
Log2 TMT ratios to the common reference were used as quantitative values for all proteomics features
(proteins, phosphosites, acetylsite, and ubiquitylsites). Proteomics features not fully quantified in at least
two plexes within a tissue and non-rat contaminants were removed. Log2 TMT ratios were sample-normalized by
median-centering and mean absolute deviation scaling. Plex batch effects were removed using linear models
implemented by the \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.
}
\keyword{datasets}
