% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHOSPHO_SAMPLE_DATA}
\alias{PHOSPHO_SAMPLE_DATA}
\title{Protein phosphorylation normalized sample-level data}
\format{
A data frame with 268037 rows and 61 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-untargeted/prot-ph/normalized-data/*med-mad-normalized-logratio.txt}
}
\usage{
PHOSPHO_SAMPLE_DATA
}
\description{
Median-MAD normalized sample-level protein phosphorylation data. Not adjusted for protein abundance.
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
