% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PROT_FEATURE_ANNOT}
\alias{PROT_FEATURE_ANNOT}
\title{Global protein expression feature annotation}
\format{
A data frame with 1379979 rows and 8 variables:
\describe{
\item{\code{protein_id}}{character, RefSeq identifier in format \code{[accession].[version]},
e.g., NP_001004415.1. Note: contaminants should be identified with the following nomenclature: "Contaminant_XXXX"}
\item{\code{redundant_ids}}{character, pipe-separated list of additional RefSeq identifiers with
redundant peptide-matching sequences to \code{protein_id} that cannot be resolved by inference due
to lack of unique peptides}
\item{\code{is_contaminant}}{logical, whether \code{protein_id} is a contaminant}
\item{\code{peptide_score}}{double, MSGF+ SpecE Value or Spectrum Mill score (\code{1/bestScore})}
\item{\code{sequence}}{character, peptide sequence determined by MSGF+ or Spectrum Mill analysis of LC-MS/MS features}
\item{\code{organism_name}}{character, organism whose protein collection database was
used for LC-MS/MS peptide identification, e.g., Rattus norvegicus}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{assay}}{character, assay "PROT"}
}
}
\source{
\code{pass1b-06/results/proteomics-untargeted/*/prot-pr/*rii-results.txt}
}
\description{
Global protein expression feature annotation
}
\details{
PROT feature annotation is only available via download from Google Cloud Storage:
\url{https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/feature-annot/PROT_FEATURE_ANNOT.rda}.
You can use \code{\link[MotrpacRatTraining6mo:load_feature_annotation]{MotrpacRatTraining6mo::load_feature_annotation()}} to download and return this file.
}
