% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_NORM_DATA}
\alias{PROT_NORM_DATA}
\alias{PROT_CORTEX_NORM_DATA}
\alias{PROT_SKMGN_NORM_DATA}
\alias{PROT_HEART_NORM_DATA}
\alias{PROT_KIDNEY_NORM_DATA}
\alias{PROT_LUNG_NORM_DATA}
\alias{PROT_LIVER_NORM_DATA}
\alias{PROT_WATSC_NORM_DATA}
\title{Normalized protein expression data}
\format{
A data frame with protein IDs in rows (\code{feature_ID}) and samples in columns (\code{viallabel})

An object of class \code{data.frame} with 5999 rows and 58 columns.

An object of class \code{data.frame} with 9184 rows and 61 columns.

An object of class \code{data.frame} with 9852 rows and 61 columns.

An object of class \code{data.frame} with 10877 rows and 61 columns.

An object of class \code{data.frame} with 9791 rows and 60 columns.

An object of class \code{data.frame} with 9964 rows and 61 columns.
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-untargeted/prot-pr/normalized-data/*med-mad-normalized-logratio.txt}
}
\usage{
PROT_CORTEX_NORM_DATA

PROT_SKMGN_NORM_DATA

PROT_HEART_NORM_DATA

PROT_KIDNEY_NORM_DATA

PROT_LUNG_NORM_DATA

PROT_LIVER_NORM_DATA

PROT_WATSC_NORM_DATA
}
\description{
Median-MAD normalized protein expression (PROT) data used for visualization and differential analysis
}
\details{
#' @title Normalized ATAC-seq data
#' @description Normalized sample-level ATAC-seq (ATAC) data used for visualization and differential analysis
#' @format A data frame with peaks in rows (\code{feature_ID}) and samples in columns (\code{viallabel})
#' @details Data was processed with the \href{https://github.com/ENCODE-DCC/atac-seq-pipeline}{ENCODE ATAC-seq pipeline (v1.7.0)}.
#'   Samples from a single sex and training time point, e.g., males trained for 2 weeks, were analyzed together as biological
#'   replicates in a single workflow. Briefly, adapters were trimmed with cutadapt v2.5 (Martin, 2011) and aligned to release 96
#'   of the Ensembl Rattus norvegicus (rn6) genome (Dobin et al., 2013) with Bowtie 2 v2.3.4.3 (Langmead and Salzberg, 2012).
#'   Duplicate reads and reads mapping to the mitochondrial chromosome were removed. Signal files and peak calls were generated
#'   using MACS2 v2.2.4 (Gaspar, 2018), both from reads from each sample and pooled reads from all biological replicates.
#'   Pooled peaks were compared with the peaks called for each replicate individually using Irreproducibility Discovery Rate (Li et al., 2011)
#'   and thresholded to generate an optimal set of peaks.
#'
#'   The cloud implementation of the ENCODE ATAC-seq pipeline and source code for the post-processing steps are available at \url{https://github.com/MoTrPAC/motrpac-atac-seq-pipeline}.
#'   Optimal peaks (overlap.optimal_peak.narrowPeak.bed.gz) from all workflows were concatenated, trimmed to 200 base pairs around the summit,
#'   and sorted and merged with bedtools v2.29.0 (Quinlan and Hall, 2010) to generate a master peak list. This peak list was intersected with
#'   the filtered alignments from each sample using bedtools coverage with options \code{-nonamecheck} and \code{-counts} to generate a peak by
#'   sample matrix of raw counts.
#'
#'   The remaining steps were applied separately on raw counts from each tissue. Peaks from non-autosomal chromosomes were removed,
#'   as well as peaks that did not have at least 10 read counts in four samples. Filtered raw counts were then quantile-normalized with
#'   limma-voom (Law et al., 2014). This version of the normalized data is provided by this object.
#'
#' @source \code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/epigenomics/epigen-atac-seq/normalized-data/*quant-norm*}
#' @name ATAC_NORM_DATA
"ATAC_HIPPOC_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_SKMGN_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_HEART_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_KIDNEY_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_LUNG_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_LIVER_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_BAT_NORM_DATA"

#' @rdname ATAC_NORM_DATA
"ATAC_WATSC_NORM_DATA"

#' @title Normalized DNA methylation data
#' @description Normalized DNA methylation (METHYL) data used for visualization
#' @format A data frame with CpG sites in rows (\code{feature_ID}) and samples in columns (\code{viallabel})
#' @details Only CpG sites with methylation coverage of >=10 in all samples were included for downstream analysis,
#'   and normalization was performed separately in each tissue. Individual CpG sites were divided into 500 base-pair
#'   windows and were clustered using the Markov Clustering algorithm via the MCL R package (Jager, 2015). To apply MCL,
#'   for each 500 base-pair window an undirected graph was constructed, linking individual sites if their correlation
#'   was >=0.7. MCL was chosen for this task as it: (1) determines the number of clusters internally, (2) identifies
#'   homogeneous clusters, and (3) keeps single sites that are not correlated with either sites as singletons (clusters of size one).
#'   The resulting sites/clusters were used as input for normalization and differential analysis with edgeR (Robinson et al., 2010).
#'   To generate this normalized sample-level data, the methylation coverages of filtered sites/clusters were first log2-transformed,
#'   and normalization was performed using \code{\link[preprocessCore:normalize.quantiles.robust]{preprocessCore::normalize.quantiles.robust()}} (Bolstad, 2021).
#'
#'   In order to stay under the 50MB GitHub file limit, data frames with more than 500k rows were split in half and saved in two
#'   files, e.g. \link{METHYL_BAT_NORM_DATA_P1} and \link{METHYL_BAT_NORM_DATA_P2}
#' @source \code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/epigenomics/epigen-rrbs/normalized-data/*normalized-log-M-window.txt}
#' @name METHYL_NORM_DATA
"METHYL_BAT_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_BAT_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_HEART_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_HEART_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_HIPPOC_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_HIPPOC_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_KIDNEY_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_KIDNEY_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_LIVER_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_LIVER_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_LUNG_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_LUNG_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_SKMGN_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_SKMGN_NORM_DATA_P2"

#' @rdname METHYL_NORM_DATA
"METHYL_WATSC_NORM_DATA_P1"

#' @rdname METHYL_NORM_DATA
"METHYL_WATSC_NORM_DATA_P2"

Log2 TMT ratios to the common reference were used as quantitative values for all proteomics features
(proteins, phosphosites, acetylsite, and ubiquitylsites). Proteomics features not fully quantified in at least
two plexes within a tissue and non-rat contaminants were removed. Log2 TMT ratios were sample-normalized by
median-centering and mean absolute deviation scaling. Plex batch effects were removed using linear models
implemented by the \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.
}
\keyword{datasets}
