% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROTEOME_DA}
\alias{PROTEOME_DA}
\alias{PROT_CORTEX_DA}
\alias{PROT_SKMGN_DA}
\alias{PROT_HEART_DA}
\alias{PROT_KIDNEY_DA}
\alias{PROT_LUNG_DA}
\alias{PROT_LIVER_DA}
\alias{PROT_WATSC_DA}
\alias{PHOSPHO_CORTEX_DA}
\alias{PHOSPHO_SKMGN_DA}
\alias{PHOSPHO_HEART_DA}
\alias{PHOSPHO_KIDNEY_DA}
\alias{PHOSPHO_LUNG_DA}
\alias{PHOSPHO_LIVER_DA}
\alias{PHOSPHO_WATSC_DA}
\alias{ACETYL_HEART_DA}
\alias{ACETYL_LIVER_DA}
\alias{UBIQ_HEART_DA}
\alias{UBIQ_LIVER_DA}
\title{Differential analysis of proteomics datasets}
\format{
A data frame 18 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{sex}}{@eval sex()}
\item{\code{comparison_group}}{@eval comparison_group()}
\item{\code{p_value}}{@eval p_value_da()}
\item{\code{adj_p_value}}{@eval adj_p_value_da()}
\item{\code{logFC}}{@eval logFC()}
\item{\code{logFC_se}}{@eval logFC_se()}
\item{\code{tscore}}{@eval tscore()}
\item{\code{covariates}}{@eval covariates()}
\item{\code{numNAs}}{@eval numNAs()}
\item{\code{comparison_average_intensity}}{@eval comparison_average_intensity()}
\item{\code{reference_average_intensity}}{@eval reference_average_intensity()}
\item{\code{selection_fdr}}{@eval selection_fdr()}
}

An object of class \code{data.frame} with 47992 rows and 18 columns.

An object of class \code{data.frame} with 73472 rows and 18 columns.

An object of class \code{data.frame} with 78816 rows and 18 columns.

An object of class \code{data.frame} with 87016 rows and 18 columns.

An object of class \code{data.frame} with 78328 rows and 18 columns.

An object of class \code{data.frame} with 79712 rows and 18 columns.

An object of class \code{data.frame} with 372016 rows and 18 columns.

An object of class \code{data.frame} with 237208 rows and 18 columns.

An object of class \code{data.frame} with 321664 rows and 18 columns.

An object of class \code{data.frame} with 241152 rows and 18 columns.

An object of class \code{data.frame} with 381288 rows and 18 columns.

An object of class \code{data.frame} with 348536 rows and 18 columns.

An object of class \code{data.frame} with 242432 rows and 18 columns.

An object of class \code{data.frame} with 41704 rows and 18 columns.

An object of class \code{data.frame} with 78000 rows and 18 columns.

An object of class \code{data.frame} with 56624 rows and 18 columns.

An object of class \code{data.frame} with 74752 rows and 18 columns.
}
\usage{
PROT_CORTEX_DA

PROT_SKMGN_DA

PROT_HEART_DA

PROT_KIDNEY_DA

PROT_LUNG_DA

PROT_LIVER_DA

PROT_WATSC_DA

PHOSPHO_CORTEX_DA

PHOSPHO_SKMGN_DA

PHOSPHO_HEART_DA

PHOSPHO_KIDNEY_DA

PHOSPHO_LUNG_DA

PHOSPHO_LIVER_DA

PHOSPHO_WATSC_DA

ACETYL_HEART_DA

ACETYL_LIVER_DA

UBIQ_HEART_DA

UBIQ_LIVER_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
proteomics feature (protein, phosphosite, acetylsite, ubiquitlsite)
within each sex. One data frame per data type per tissue.
}
\keyword{datasets}
