% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROTEOMICS_META}
\alias{PROTEOMICS_META}
\alias{PHOSPHO_META}
\alias{PROT_META}
\alias{ACETYL_META}
\alias{UBIQ_META}
\title{Proteomics sample-level metadata}
\format{
A data frame with 5 variables:
\describe{
\item{\code{tmt11_channel}}{character, TMT channel}
\item{\code{tmt_plex}}{character, TMT plex}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{viallabel}}{character, sample identifier}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
}

An object of class \code{data.frame} with 462 rows and 5 columns.

An object of class \code{data.frame} with 120 rows and 5 columns.

An object of class \code{data.frame} with 132 rows and 5 columns.
}
\usage{
PHOSPHO_META

PROT_META

ACETYL_META

UBIQ_META
}
\description{
TMT channel and plex information for each proteomics assay
}
\details{
TODO
}
\keyword{datasets}
