% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAT_TO_HUMAN_GENE}
\alias{RAT_TO_HUMAN_GENE}
\title{Rat-to-human gene ortholog map}
\format{
A data frame with 21461 rows and 12 variables:
\describe{
\item{\code{RAT_SYMBOL}}{character, official rat gene symbol}
\item{\code{NAME}}{character, gene name}
\item{\code{RAT_GENE_RGD_ID}}{integer, rat RGD gene ID}
\item{\code{RAT_NCBI_GENE_ID}}{integer, rat NCBI/Entrez gene ID}
\item{\code{RAT_ENSEMBL_ID}}{character, rat Ensembl gene ID}
\item{\code{RAT_UNIPROT_ID}}{character, semicolon-separated list of rat UniProt gene IDs}
\item{\code{RAT_OLD_SYMBOL}}{character, semicolon-separated list of rat deprecated or alias gene symbols}
\item{\code{HUMAN_ORTHOLOG_SYMBOL}}{character, human official gene symbol}
\item{\code{HUMAN_ORTHOLOG_RGD}}{character, human RGD gene ID}
\item{\code{HUMAN_ORTHOLOG_NCBI_GENE_ID}}{character, human NCBI/Entrez gene ID}
\item{\code{HUMAN_ORTHOLOG_ENSEMBL_ID}}{character, human Entrez gene ID}
\item{\code{HUMAN_ORTHOLOG_SOURCE}}{character, human ortholog source}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{<https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_39/gencode.v39.metadata.EntrezGene.gz>
<https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_39/gencode.v39.metadata.HGNC.gz>
<https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_39/gencode.v39.chr_patch_hapl_scaff.basic.annotation.gtf.gz>
<https://download.rgd.mcw.edu/pub/data_release/RGD_ORTHOLOGS.txt> (2020-01-10)
<https://download.rgd.mcw.edu/data_release/RAT/GENES_RAT.txt> (2020-11-15)
}\if{html}{\out{</div>}}
}
\usage{
RAT_TO_HUMAN_GENE
}
\description{
Mapping between various rat and human gene identifiers for orthologs as reported by \href{https://rgd.mcw.edu/}{RGD} (2020-01-10).
}
\details{
This map was compiled from several external sources.
GENCODE metadata and annotation files were used to map between human Ensembl transcript IDs, Entrez IDs,
GENCODE IDs, and Ensembl gene IDs \href{https://pubmed.ncbi.nlm.nih.gov/33270111/}{(Frankish et al., 2021)}. RGD files were used to map between human and
rat gene symbols as well as between various rat gene identifiers \href{https://pubmed.ncbi.nlm.nih.gov/31713623/}{(Smith et al., 2020)}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{There is one row per unique combination of \code{RAT_SYMBOL}, \code{RAT_GENE_RGD_ID}, \code{RAT_NCBI_GENE_ID}, \code{RAT_ENSEMBL_ID},
\code{RAT_UNIPROT_ID}, \code{HUMAN_ORTHOLOG_SYMBOL}, \code{HUMAN_ORTHOLOG_RGD}, \code{HUMAN_ORTHOLOG_NCBI_GENE_ID}, and \code{HUMAN_ORTHOLOG_ENSEMBL_ID}, 
so some genes correspond to multiple rows in the table. For example, there are 24 rows corresponding to \code{RAT_SYMBOL == "Tnf"} 
because 3 \code{RAT_ENSEMBL_ID}s correspond to Tnf, and 8 \code{HUMAN_ORTHOLOG_ENSEMBL_ID}s correspond to TNF.  
}\if{html}{\out{</div>}}
}
\keyword{datasets}
