% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REPFDR_INPUTS}
\alias{REPFDR_INPUTS}
\title{\code{repfdr} inputs}
\format{
List of data frames:
\describe{
\item{\code{zs_info}}{feature-level metadata for training-regulated features included in graphical analysis}
\item{\code{zs_smoothed}}{feature by state data frame where values are smoothed timewise z-scores}
\item{\code{nominal_ps}}{feature by state data frame where values are timewise nominal p-values}
\item{\code{effects}}{feature by state data frame where values are timewise effects or log fold-changes}
}
}
\usage{
REPFDR_INPUTS
}
\description{
Feature by state data frames that define the input for \code{\link[repfdr:repfdr]{repfdr::repfdr()}}
}
\keyword{datasets}
