% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEX_COLORS}
\alias{SEX_COLORS}
\title{Sex colors}
\format{
Named character vector
}
\usage{
SEX_COLORS
}
\description{
Sex colors used in figures. Values are hex codes and names include multiple ways to indicate male or female.
}
\keyword{datasets}
