% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TISSUE_ABBREV}
\alias{TISSUE_ABBREV}
\title{Tissue abbreviations}
\format{
Unnamed character vector
}
\usage{
TISSUE_ABBREV
}
\description{
Tissue abbreviations used in tables and figures
}
\keyword{datasets}
