% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TISSUE_ABBREV_TO_CODE}
\alias{TISSUE_ABBREV_TO_CODE}
\title{Tissue abbreviation-to-code mapping}
\format{
Named character vector
}
\usage{
TISSUE_ABBREV_TO_CODE
}
\description{
Values are codes and names are abbreviations (\code{TISSUE_ABBREV}).
See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code} for more details.
}
\keyword{datasets}
