% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TISSUE_COLORS}
\alias{TISSUE_COLORS}
\title{Tissue colors}
\format{
Named character vector
}
\usage{
TISSUE_COLORS
}
\description{
Tissue colors used in figures. Values are hex codes and names are tissue abbreviations \link{TISSUE_ABBREV} OR tissue codes.
}
\keyword{datasets}
