% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TISSUE_ORDER}
\alias{TISSUE_ORDER}
\title{Tissue order}
\format{
Unnamed character vector
}
\usage{
TISSUE_ORDER
}
\description{
Biological order of \link{TISSUE_ABBREV} used in figures
}
\keyword{datasets}
