% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TRAINING_DA}
\alias{TRAINING_DA}
\title{Training differential analysis results}
\format{
A data frame with up to 47 variables:
\describe{
\item{\code{assay}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{removed_samples}}{character, comma-separated list of vial labels excluded from differential analysis or NA}
\item{\code{removed_samples_male}}{character, comma-separated list of male outliers (vial labels) removed from differential analysis}
\item{\code{removed_samples_female}}{character, comma-separated list of female outliers (vial labels) removed from differential analysis}
\item{\code{dataset}}{character, specific dataset}
\item{\code{dataset_male}}{character, specific dataset}
\item{\code{dataset_female}}{character, specific dataset}
\item{\code{panel}}{character, LUMINEX panel for immunoassay analytes}
\item{\code{full_model}}{character, full model used for differential analysis}
\item{\code{full_model_male}}{character, full model used for differential analysis in males}
\item{\code{full_model_female}}{character, full model used for differential analysis in females}
\item{\code{reduced_model}}{character, effective reduced model used for differential analysis}
\item{\code{reduced_model_male}}{character, effective reduced model used for differential analysis in males}
\item{\code{reduced_model_female}}{character, effective reduced model used for differential analysis in females}
\item{\code{covariates}}{character, comma-separated list of adjustment variables}
\item{\code{covariates_male}}{character, comma-separated list of adjustment variables used for males}
\item{\code{covariates_female}}{character, comma-separated list of adjustment variables used for females}
\item{\code{fscore_male}}{double, F-statistic for males}
\item{\code{fscore_female}}{double, F-statistic for females}
\item{\code{lrt}}{double, likelihood ratio test statistic when only one sex exists}
\item{\code{lrt_male}}{double, likelihood ratio test statistic for males}
\item{\code{lrt_female}}{double, likelihood ratio test statistic for females}
\item{\code{p_value_male}}{double, nominal p-value for males}
\item{\code{p_value_female}}{character, nominal p-value for females}
\item{\code{adj_p_value_male}}{double, IHW-adjusted p-value for males}
\item{\code{adj_p_value_female}}{double, IHW-adjusted p-value for females}
\item{\code{p_value}}{double, combined male and female nominal p-value using the sum of logs}
\item{\code{adj_p_value}}{double, IHW-adjusted combined p-value; value provided by \code{selection_fdr} in the timewise summary statistics}
\item{\code{tissue_abbreviation}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{Chr}}{character, chromosome where feature is located. METHYL only.}
\item{\code{Locus}}{character, name of feature. METHYL only.}
\item{\code{EntrezID}}{integer, Entrez ID of corresponding gene. METHYL only.}
\item{\code{Symbol}}{character, gene symbol of corresponding gene. METHYL only.}
\item{\code{is_targeted}}{logical, whether metabolite is targeted. Metabolomics only.}
\item{\code{site}}{character, Chemical Analysis Site (CAS) name. Metabolomics only.}
\item{\code{metabolite}}{character, name of metabolite as appears in the CAS's data. Metabolomics only.}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite. Metabolomics only.}
\item{\code{groups_tested_male}}{character, timepoints used to perform the F-test in males.
Some tissues or assays are missing timepoints. Metabolomics only.}
\item{\code{groups_tested_female}}{character, timepoints used to perform the F-test in females.
Some tissues or assays are missing timepoints. Metabolomics only.}
\item{\code{mz}}{double, mass over charge. Metabolomics only.}
\item{\code{rt}}{double, retention time. Metabolomics only.}
\item{\code{neutral_mass}}{numeric, neutral mass. Metabolomics only.}
\item{\code{cv}}{double, feature coffeicient of variation in the dataset. Metabolomics only.}
\item{\code{meta_reg_het_p}}{double, heterogeneity p-value from meta-regression. A smaller p-value indicates more disagreement between platforms.
Metabolomics meta-regression only.}
\item{\code{original_ftest_ps}}{character, comma-separated list of original F-test p-values. Metabolomics meta-regression only.}
}
}
\description{
Training summary statistics from differential analysis (DA)
that tests the effect of training on each feature within each sex.
One data frame per data type and tissue.
}
\details{
While the more commonly-used timewise summary statistics are available directly in
this package (e.g., \link{TRNSCRPT_DA}, \link{PROTEOME_DA}, \link{IMMUNO_DA}, \link{METAB_DA}, \link{METAB_DA_METAREG}, \link{TRAINING_REGULATED_FEATURES}, \link{REPFDR_INPUTS}),
training summary statistics are available for download from Google Cloud Storage.
Use \code{\link[MotrpacRatTraining6mo:load_training_da]{MotrpacRatTraining6mo::load_training_da()}} to download results for a given tissue and ome.
Note that the adjusted p-value from the training differential analysis is already included in the
timewise summary statistics as \code{selection_fdr}.

Training differential analysis was performed using either F-tests or likelihood
ratio tests depending on the ome. Find the functions used to generate these results listed
\href{https://motrpac.github.io/MotrpacRatTraining6mo/reference/index.html#perform-differential-analysis}{here}.
See details in the supplementary methods of the \href{https://github.com/MoTrPAC/MotrpacRatTraining6moData#citing-motrpac-data}{manuscript}.
}
