% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRAINING_REGULATED_FEATURES}
\alias{TRAINING_REGULATED_FEATURES}
\title{Training-regulated features}
\format{
A data frame with 279002 rows and 18 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{non_redundant_feature_ID}}{character, non-redundant feature identifier defined for \code{feature_ID}s
with measurements from multiple platforms. \code{feature} uses \code{non_redundant_feature_ID}
when available and \code{feature_ID} otherwise. See \link{REPEATED_FEATURES} for more details.}
\item{\code{platform}}{@eval dataset()}
\item{\code{sex}}{@eval sex()}
\item{\code{training_group}}{character, training time point in weeks, one of "1w", "2w", "4w", "8w".
Corresponds to \code{comparison_group} in the \code{*_DA} tables.}
\item{\code{timewise_logFC}}{double, log fold-change of the training group specified by
\code{sex} and \code{training_group} (e.g., 1-week females) relative to
the sex-matched sedentary controls. Corresponds to \code{logFC} in the \code{*_DA} tables.}
\item{\code{timewise_logFC_se}}{double, standard error of \code{timewise_logFC}.
Corresponds to \code{logFC_se} in the \code{*_DA} tables.}
\item{\code{timewise_p_value}}{double, timewise p-value corresponding to the
contrast between the training group (e.g., 1-week females) and the sex-matched
sedentary controls. Corresponds to \code{p_value} in the \code{*_DA} tables.}
\item{\code{timewise_zscore}}{double, smoothed z-score corresponding to
\code{timewise_p_value} used for graphical clustering.
Corresponds to \code{zscore} or \code{tscore} in the \code{*_DA} tables.}
\item{\code{meta_reg_het_p}}{@eval meta_reg_het_p()}
\item{\code{meta_reg_pvalue}}{@eval meta_reg_pvalue()}
\item{\code{training_p_value}}{double, combined training p-value determined from
F-tests or LRTs comparing a full model with ome-specific technical covariates
and training group as a factor variable against a reduced model with only
technical covariates. Sex-specific training p-values were merged using the
sum of logs. One value per features (not per training group)}
\item{\code{training_q}}{double, IHW FDR-adjusted \code{training_p_value};
training-regulated features were defined by \code{training_q < 0.05}.
Corresponds to \code{selection_fdr} in the \code{*_DA} tables.}
}
}
\usage{
TRAINING_REGULATED_FEATURES
}
\description{
Differential analysis results for training-regulated features at 5\% FDR
}
\keyword{datasets}
