% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRAINING_REGULATED_NORM_DATA}
\alias{TRAINING_REGULATED_NORM_DATA}
\alias{TRAINING_REGULATED_NORM_DATA_NO_OUTLIERS}
\title{Normalized data for training-regulated features}
\format{
A data frame with features in rows and participant IDs (PIDs) in columns
(in addition to feature metadata columns)

An object of class \code{data.frame} with 35872 rows and 76 columns.
}
\usage{
TRAINING_REGULATED_NORM_DATA

TRAINING_REGULATED_NORM_DATA_NO_OUTLIERS
}
\description{
Combined normalized sample-level data for all training-regulated features
at 5\% FDR (see \link{TRAINING_REGULATED_FEATURES}). Specifically used to speed up plotting
of training-regulated features in \code{\link[MotrpacRatTraining6mo:plot_feature_trajectories]{MotrpacRatTraining6mo::plot_feature_trajectories()}}.
}
\details{
Sample-level data were compiled using \code{\link[MotrpacRatTraining6mo:load_sample_data]{MotrpacRatTraining6mo::load_sample_data()}}
with \code{normalized=TRUE, training_regulated_only=TRUE}. \link{TRAINING_REGULATED_NORM_DATA_NO_OUTLIERS}
excludes outliers defined in \link{OUTLIERS} while \link{TRAINING_REGULATED_NORM_DATA} does not.
}
\keyword{datasets}
