% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TRNSCRPT_FEATURE_ANNOT}
\alias{TRNSCRPT_FEATURE_ANNOT}
\title{Transcriptomics gene annotation}
\format{
A data frame with 32883 rows and 13 variables:
\describe{
\item{\code{seqname}}{character, name of the chromosome or scaffold. Chromosome names have the 'chr' prefix.}
\item{\code{source}}{character, name of the program that generated this feature, or the data source (database or project name)}
\item{\code{feature}}{character, feature type name, e.g. Gene, Variation, Similarity}
\item{\code{start}}{integer, start position of the feature, with sequence numbering starting at 1}
\item{\code{end}}{integer, end position of the feature, with sequence numbering starting at 1}
\item{\code{score}}{character, a floating point value}
\item{\code{strand}}{character, defined as + (forward) or - (reverse)}
\item{\code{frame}}{character, one of '0', '1' or '2'.
'0' indicates that the first base of the feature is the first base of a codon,
'1' that the second base is the first base of a codon, and so on.}
\item{\code{gene_id}}{character, Ensembl gene ID from the 'attribute' field of the GTF}
\item{\code{gene_version}}{character, gene version from the 'attribute' field of the GTF}
\item{\code{gene_name}}{character, gene name from the 'attribute' field of the GTF}
\item{\code{gene_source}}{character, gene source from the 'attribute' field of the GTF}
\item{\code{gene_biotype}}{character, gene biotype from the 'attribute' field of the GTF}
}
}
\source{
\url{https://ftp.ensembl.org/pub/release-95/gtf/rattus_norvegicus/Rattus_norvegicus.Rnor_6.0.95.gtf.gz}
}
\description{
Gene annotation adapted from Ensembl's \emph{Rattus norvegicus} rn6 release 95 GTF.
Importantly, MoTrPAC prepended autosomes 'chr' and changed 'MT' to 'chrM'.
}
\details{
TRNSCRPT feature annotation is only available via download from Google Cloud Storage:
\url{https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/feature-annot/TRNSCRPT_FEATURE_ANNOT.rda}.
You can use \code{\link[MotrpacRatTraining6mo:load_feature_annotation]{MotrpacRatTraining6mo::load_feature_annotation()}} to download and return this file.
}
