% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_META}
\alias{TRNSCRPT_META}
\title{RNA-seq metadata and QC}
\format{
A data frame with 935 rows and 82 variables:
\describe{
\item{\code{viallabel}}{character, sample identifier}
\item{\code{vial_label}}{double, sample identifier}
\item{\code{2D_barcode}}{double, sample barcode}
\item{\code{Species}}{character, species}
\item{\code{BID}}{integer, biospecimen ID}
\item{\code{PID}}{double, participant ID, one per animal}
\item{\code{Tissue}}{character, tissue description}
\item{\code{Sample_category}}{character, study sample ("study") or reference standard ("ref)}
\item{\code{GET_site}}{character, which Genomics, Epigenomics, and Transcriptomics (GET)
site performed the assay, "Stanford" or "MSSM" (Icahn School of Medicine at Mount Sinai)}
\item{\code{RNA_extr_plate_ID}}{character, RNA extraction plate ID}
\item{\code{RNA_extr_date}}{character, RNA extraction date}
\item{\code{RNA_extr_conc}}{double, RNA concentration (ng/uL)}
\item{\code{RIN}}{double, RNA Integrity Number}
\item{\code{r_260_280}}{double, 260/280 ratio}
\item{\code{r_260_230}}{double 260/230 ratio}
\item{\code{Lib_prep_date}}{character, library preparation date in MM/DD/YYYY format}
\item{\code{Lib_RNA_conc}}{double, RNA concentration used for library prep (ng/uL)}
\item{\code{Lib_RNA_vol}}{integer, RNA volume used for library prep (uL)}
\item{\code{Lib_robot}}{character, robot used for library prep}
\item{\code{Lib_vendor}}{character, library prep vendor}
\item{\code{Lib_type}}{character, library prep type}
\item{\code{Lib_kit_id}}{character, library prep kit ID}
\item{\code{Lib_batch_ID}}{character, library prep batch ID that distinguished different sample processing batches}
\item{\code{Lib_barcode_well}}{character, well}
\item{\code{Lib_index_1}}{character, i7 index}
\item{\code{Lib_index_2}}{character, i5 index}
\item{\code{Lib_adapter_1}}{character, Truseq I7 index with 16bp index}
\item{\code{Lib_adapter_2}}{character, Truseq I5 index with 8bp index}
\item{\code{Lib_UMI_cycle_num}}{integer, number of bases of UMI}
\item{\code{Lib_adapter_size}}{integer, total size of the two adapters}
\item{\code{Lib_frag_size}}{integer, average library fragment size (bp)}
\item{\code{Lib_DNA_conc}}{double, DNA concentration of original stock of the library (ng/uL)}
\item{\code{Lib_molarity}}{double, library molarity (nM)}
\item{\code{Seq_platform}}{character, sequencing platform}
\item{\code{Seq_date}}{integer, sequencing date, YYMMDD format}
\item{\code{Seq_machine_ID}}{character, serial number of the sequencer}
\item{\code{Seq_flowcell_ID}}{character, flow cell ID}
\item{\code{Seq_flowcell_run}}{integer, flow cell run}
\item{\code{Seq_flowcell_lane}}{character, flow cell lane}
\item{\code{Seq_flowcell_type}}{character, flow cell type, e.g., S4}
\item{\code{Seq_length}}{integer, read length}
\item{\code{Seq_end_type}}{integer, 1=single-end, 2=paired-end}
\item{\code{Phase}}{character, study phase, "PASS1B-06"}
\item{\code{Seq_batch}}{character, unique identifier for sequencing batch}
\item{\code{reads_raw}}{double, number of read pairs in the raw FASTQ}
\item{\code{pct_adapter_detected}}{double, percent of reads with adapter detected}
\item{\code{pct_trimmed}}{double, percent of reads that were trimmed}
\item{\code{pct_trimmed_bases}}{double, percent of bases that were trimmed}
\item{\code{reads}}{double, number of read pairs in the trimmed FASTQ files}
\item{\code{pct_GC}}{double, percent GC content in trimmed FASTQ files}
\item{\code{pct_dup_sequence}}{double, percent of duplicated sequences in trimmed FASTQ files}
\item{\code{pct_rRNA}}{double, percent of rRNA reads in trimmed FASTQ files}
\item{\code{pct_globin}}{double, percent of globin reads in trimmed FASTQ files}
\item{\code{pct_phix}}{double, percent of phix reads in trimmed FASTQ files}
\item{\code{pct_picard_dup}}{double, PCR duplication assessed by Picard’s tool MarkDuplicate}
\item{\code{pct_umi_dup}}{double, PCR duplication rate assessed using UMIs (Unique Molecular Identifiers)}
\item{\code{avg_input_read_length}}{double, average input read length}
\item{\code{uniquely_mapped}}{double, number of uniquely mapped reads}
\item{\code{pct_uniquely_mapped}}{double, percent of uniquely mapped reads}
\item{\code{avg_mapped_read_length}}{double, average input mapped length}
\item{\code{num_splices}}{double, number of splices}
\item{\code{num_annotated_splices}}{double, number of annotated splices}
\item{\code{num_GTAG_splices}}{double, number of GT/AG and CT/AC splices}
\item{\code{num_GCAG_splices}}{double, number of GC/AG and CT/GC splices}
\item{\code{num_ATAC_splices}}{double, number of AT/AC and GT/TA splices}
\item{\code{num_noncanonical_splices}}{double, number of non-canonical splices}
\item{\code{pct_multimapped}}{double, percent of reads that multimapped}
\item{\code{pct_multimapped_toomany}}{double, percent of reads that multimapped too many times}
\item{\code{pct_unmapped_mismatches}}{double, percent of unmapped reads due to mismatches}
\item{\code{pct_unmapped_tooshort}}{double, percent of unmapped reads due to shortness}
\item{\code{pct_unmapped_other}}{double, percent of unmapped reads for other reason}
\item{\code{pct_chimeric}}{double, percent chimeric reads}
\item{\code{pct_chrX}}{double, percent of reads mapped to chromosome X}
\item{\code{pct_chrY}}{double, percent of reads mapped to chromosome Y}
\item{\code{pct_chrM}}{double, percent of reads mapped to the mitochondrial genome}
\item{\code{pct_chrAuto}}{double, percent of reads mapped to autosomal chromosomes}
\item{\code{pct_contig}}{double, percent of reads mapped to contigs}
\item{\code{pct_coding}}{double, percent of bases mapped to coding}
\item{\code{pct_utr}}{double, percent of bases mapped to untranslated region}
\item{\code{pct_intronic}}{double, percent of bases mapped to introns}
\item{\code{pct_intergenic}}{double, percent of bases mapped to intergenic}
\item{\code{pct_mrna}}{double, percent of bases mapped to mRNA}
\item{\code{median_5_3_bias}}{double, median 5' to 3' bias}
}
}
\source{
\code{pass1b-06/results/transcriptomics/qa-qc/motrpac_pass1b-06_transcript-rna-seq_qa-qc-metrics.csv}
}
\usage{
TRNSCRPT_META
}
\description{
RNA-seq experimental and quantification QC metrics for transcriptomic (TRNSCRPT) data
}
\keyword{datasets}
