% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_NORM_DATA}
\alias{TRNSCRPT_NORM_DATA}
\alias{TRNSCRPT_BLOOD_NORM_DATA}
\alias{TRNSCRPT_HIPPOC_NORM_DATA}
\alias{TRNSCRPT_CORTEX_NORM_DATA}
\alias{TRNSCRPT_HYPOTH_NORM_DATA}
\alias{TRNSCRPT_SKMGN_NORM_DATA}
\alias{TRNSCRPT_SKMVL_NORM_DATA}
\alias{TRNSCRPT_HEART_NORM_DATA}
\alias{TRNSCRPT_KIDNEY_NORM_DATA}
\alias{TRNSCRPT_ADRNL_NORM_DATA}
\alias{TRNSCRPT_COLON_NORM_DATA}
\alias{TRNSCRPT_SPLEEN_NORM_DATA}
\alias{TRNSCRPT_TESTES_NORM_DATA}
\alias{TRNSCRPT_OVARY_NORM_DATA}
\alias{TRNSCRPT_VENACV_NORM_DATA}
\alias{TRNSCRPT_LUNG_NORM_DATA}
\alias{TRNSCRPT_SMLINT_NORM_DATA}
\alias{TRNSCRPT_LIVER_NORM_DATA}
\alias{TRNSCRPT_BAT_NORM_DATA}
\alias{TRNSCRPT_WATSC_NORM_DATA}
\title{Normalized RNA-seq data}
\format{
A data frame with genes in rows (\code{feature_ID}) and samples in columns (\code{viallabel})

An object of class \code{data.frame} with 16608 rows and 54 columns.

An object of class \code{data.frame} with 16443 rows and 54 columns.

An object of class \code{data.frame} with 16955 rows and 54 columns.

An object of class \code{data.frame} with 13774 rows and 54 columns.

An object of class \code{data.frame} with 13973 rows and 54 columns.

An object of class \code{data.frame} with 14445 rows and 54 columns.

An object of class \code{data.frame} with 15981 rows and 54 columns.

An object of class \code{data.frame} with 15809 rows and 54 columns.

An object of class \code{data.frame} with 16556 rows and 54 columns.

An object of class \code{data.frame} with 16319 rows and 54 columns.

An object of class \code{data.frame} with 17359 rows and 29 columns.

An object of class \code{data.frame} with 17035 rows and 28 columns.

An object of class \code{data.frame} with 16338 rows and 54 columns.

An object of class \code{data.frame} with 16505 rows and 54 columns.

An object of class \code{data.frame} with 16411 rows and 54 columns.

An object of class \code{data.frame} with 14437 rows and 54 columns.

An object of class \code{data.frame} with 16154 rows and 54 columns.

An object of class \code{data.frame} with 16764 rows and 54 columns.
}
\source{
\code{pass1b-06/analysis/transcriptomics/transcript-rna-seq/normalized-data/*normalized-log-cpm*}
}
\usage{
TRNSCRPT_BLOOD_NORM_DATA

TRNSCRPT_HIPPOC_NORM_DATA

TRNSCRPT_CORTEX_NORM_DATA

TRNSCRPT_HYPOTH_NORM_DATA

TRNSCRPT_SKMGN_NORM_DATA

TRNSCRPT_SKMVL_NORM_DATA

TRNSCRPT_HEART_NORM_DATA

TRNSCRPT_KIDNEY_NORM_DATA

TRNSCRPT_ADRNL_NORM_DATA

TRNSCRPT_COLON_NORM_DATA

TRNSCRPT_SPLEEN_NORM_DATA

TRNSCRPT_TESTES_NORM_DATA

TRNSCRPT_OVARY_NORM_DATA

TRNSCRPT_VENACV_NORM_DATA

TRNSCRPT_LUNG_NORM_DATA

TRNSCRPT_SMLINT_NORM_DATA

TRNSCRPT_LIVER_NORM_DATA

TRNSCRPT_BAT_NORM_DATA

TRNSCRPT_WATSC_NORM_DATA
}
\description{
Normalized sample-level RNA-seq (TRNSCRPT) data used for visualization
}
\details{
Filtering of lowly expressed genes and normalization were performed
separately in each tissue. RSEM gene counts were used to remove lowly expressed genes,
defined as having 0.5 or fewer counts per million in all but one sample.
To generate normalized sample-level data, filtered gene counts were
TMM-normalized using \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}, followed by conversion to log counts per million with \code{\link[edgeR:cpm]{edgeR::cpm()}}.
}
\keyword{datasets}
