% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_RAW_COUNTS}
\alias{TRNSCRPT_RAW_COUNTS}
\alias{TRNSCRPT_ADRNL_RAW_COUNTS}
\alias{TRNSCRPT_BAT_RAW_COUNTS}
\alias{TRNSCRPT_BLOOD_RAW_COUNTS}
\alias{TRNSCRPT_COLON_RAW_COUNTS}
\alias{TRNSCRPT_CORTEX_RAW_COUNTS}
\alias{TRNSCRPT_HEART_RAW_COUNTS}
\alias{TRNSCRPT_HIPPOC_RAW_COUNTS}
\alias{TRNSCRPT_HYPOTH_RAW_COUNTS}
\alias{TRNSCRPT_KIDNEY_RAW_COUNTS}
\alias{TRNSCRPT_LIVER_RAW_COUNTS}
\alias{TRNSCRPT_LUNG_RAW_COUNTS}
\alias{TRNSCRPT_OVARY_RAW_COUNTS}
\alias{TRNSCRPT_SKMGN_RAW_COUNTS}
\alias{TRNSCRPT_SKMVL_RAW_COUNTS}
\alias{TRNSCRPT_SMLINT_RAW_COUNTS}
\alias{TRNSCRPT_SPLEEN_RAW_COUNTS}
\alias{TRNSCRPT_TESTES_RAW_COUNTS}
\alias{TRNSCRPT_VENACV_RAW_COUNTS}
\alias{TRNSCRPT_WATSC_RAW_COUNTS}
\title{RNA-seq raw counts}
\format{
A data frame with 32883 rows and one column per viallabel

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 28 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 29 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.

An object of class \code{data.frame} with 32883 rows and 54 columns.
}
\usage{
TRNSCRPT_ADRNL_RAW_COUNTS

TRNSCRPT_BAT_RAW_COUNTS

TRNSCRPT_BLOOD_RAW_COUNTS

TRNSCRPT_COLON_RAW_COUNTS

TRNSCRPT_CORTEX_RAW_COUNTS

TRNSCRPT_HEART_RAW_COUNTS

TRNSCRPT_HIPPOC_RAW_COUNTS

TRNSCRPT_HYPOTH_RAW_COUNTS

TRNSCRPT_KIDNEY_RAW_COUNTS

TRNSCRPT_LIVER_RAW_COUNTS

TRNSCRPT_LUNG_RAW_COUNTS

TRNSCRPT_OVARY_RAW_COUNTS

TRNSCRPT_SKMGN_RAW_COUNTS

TRNSCRPT_SKMVL_RAW_COUNTS

TRNSCRPT_SMLINT_RAW_COUNTS

TRNSCRPT_SPLEEN_RAW_COUNTS

TRNSCRPT_TESTES_RAW_COUNTS

TRNSCRPT_VENACV_RAW_COUNTS

TRNSCRPT_WATSC_RAW_COUNTS
}
\description{
RNA-seq raw counts as quantified by RSEM
}
\details{
STAR (v2.7.0d) was used to index and align reads to release 96
of the Ensembl Rattus norvegicus (rn6) genome and gene annotation.
RSEM (v1.3.1) was used to quantify transcriptome-coordinate-sorted
alignments using a forward probability of 0.5 to indicate a non-strand-specific protocol.
}
\keyword{datasets}
