% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_SAMPLE_DATA}
\alias{TRNSCRPT_SAMPLE_DATA}
\title{RNA-seq normalized sample-level data}
\format{
A data frame with 301801 rows and 51 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/transcriptomics/transcript-rna-seq/normalized-data/*normalized-log-cpm*}
}
\usage{
TRNSCRPT_SAMPLE_DATA
}
\description{
TMM-normalized RNA-seq counts (log CPM)
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
