% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{UBIQ_FEATURE_ANNOT}
\alias{UBIQ_FEATURE_ANNOT}
\title{Protein ubiquitination feature annotation}
\format{
A data frame with 21436 rows and 12 variables:
\describe{
\item{\code{protein_id}}{character, RefSeq identifier in format \code{[accession].[version]},
e.g., NP_001004415.1. Note: contaminants should be identified with the following nomenclature: "Contaminant_XXXX"}
\item{\code{redundant_ids}}{character, pipe-separated list of additional RefSeq identifiers with
redundant peptide-matching sequences to \code{protein_id} that cannot be resolved by inference due
to lack of unique peptides}
\item{\code{is_contaminant}}{logical, whether \code{protein_id} is a contaminant}
\item{\code{peptide_score}}{double, MSGF+ SpecE Value or Spectrum Mill score (\code{1/bestScore})}
\item{\code{sequence}}{character, peptide sequence determined by MSGF+ or Spectrum Mill analysis of LC-MS/MS features}
\item{\code{organism_name}}{character, organism whose protein collection database was
used for LC-MS/MS peptide identification, e.g., Rattus norvegicus}
\item{\code{ptm_id}}{character, concatenation of \code{protein_id}
and the modified amino acid residue (e.g., serine residue, 20 denoted as S20s),
with "-" as a separator}
\item{\code{ptm_peptide}}{character, concatenation of \code{ptm_id} and peptide with "-" as a separator.
Values in this column are unique.}
\item{\code{confident_score}}{double, score indicating confidence
of modified amino acid residue localization by LC-MS/MS fragmentation as calculated by
Ascore algorith (0-1000) or Spectrum Mill (0-1)}
\item{\code{confident_site}}{logical, whether a modified site was confidently
localized by LC-MS/MS fragmentation}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{assay}}{character, assay "UBIQ"}
}
}
\source{
\code{pass1b-06/results/proteomics-untargeted/*/prot-ub/*rii-results.txt}
}
\description{
Protein ubiquitination feature annotation
}
\details{
UBIQ feature annotation is only available via download from Google Cloud Storage:
\url{https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/feature-annot/UBIQ_FEATURE_ANNOT.rda}.
You can use \code{\link[MotrpacRatTraining6mo:load_feature_annotation]{MotrpacRatTraining6mo::load_feature_annotation()}} to download and return this file.
}
