% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UBIQ_NORM_DATA}
\alias{UBIQ_NORM_DATA}
\alias{UBIQ_HEART_NORM_DATA}
\alias{UBIQ_LIVER_NORM_DATA}
\title{Normalized protein ubiquitynation data}
\format{
A data frame with phosphosites in rows (\code{feature_ID}) and samples in columns (\code{viallabel})

An object of class \code{data.frame} with 9344 rows and 64 columns.
}
\source{
\code{pass1b-06/analysis/proteomics-untargeted/prot-ph/normalized-data/*med-mad-normalized-protein-corrected-logratio.txt}
}
\usage{
UBIQ_HEART_NORM_DATA

UBIQ_LIVER_NORM_DATA
}
\description{
Median-MAD normalized protein ubiquitynation (UBIQ) data used for visualization and differential analysis
}
\details{
Log2 TMT ratios to the common reference were used as quantitative values for all proteomics features
(proteins, phosphosites, acetylsite, and ubiquitylsites). Proteomics features not fully quantified in at least
two plexes within a tissue and non-rat contaminants were removed. Log2 TMT ratios were sample-normalized by
median-centering and mean absolute deviation scaling. Plex batch effects were removed using linear models
implemented by the \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}. The ubiquitynation datasets were corrected for changes in protein abundances
by fitting a global linear model between the ubiquitylsite and the cognate protein and extracting the residuals.
}
\keyword{datasets}
