% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UBIQ_SAMPLE_DATA}
\alias{UBIQ_SAMPLE_DATA}
\title{Protein ubiquitynation normalized sample-level data}
\format{
A data frame with 16422 rows and 61 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-untargeted/prot-ub/normalized-data/*med-mad-normalized-protein-corrected-logratio.txt}
}
\usage{
UBIQ_SAMPLE_DATA
}
\description{
Median-MAD normalized sample-level protein ubiquitynation data. Adjusted for protein abundance.
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
