% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ATAC_DA}
\alias{ATAC_DA}
\title{Differential analysis of ATAC-seq data}
\format{
TODO
}
\description{
TODO
}
\details{
TODO
Full ATAC differential analysis results are only available via download from the Cloud. See TODO.

For training-regulated ATAC features at 5\% FDR, see \link{TRAINING_REGULATED_FEATURES}.
}
