% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRAPH_STATES}
\alias{GRAPH_STATES}
\title{\code{repfdr} state assignments}
\format{
A data frame with 34244 rows and 10 variables:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link{REPEATED_FEATURES} for details.}
\item{\code{ome}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}. Note that VENACV, OVARY, TESTES,
were not included in the graphical representation of differential
features due to missing groups (e.g., females trained for 1 week).}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{state_1w}}{character, state (1, up-regulated; 0, null; -1, down-regulated)
of the feature in each sex (F, females; M, males) at the 1-week training time point,
relative to sex-matched untrained animals}
\item{\code{state_2w}}{character, state of the feature in each sex at the 2-week training time point}
\item{\code{state_4w}}{character, state of the feature in each sex at the 4-week training time point}
\item{\code{state_8w}}{character, state of the feature in each sex at the 8-week training time point}
\item{\code{path}}{character, assigned states from weeks 1-8, separated by "->".
This represents a feature's full path through the graph.
NA if the state at any of the four time points is NA.}
\item{\code{tissue_path}}{character, assigned states from weeks 1-8, separated by "->".
This represents a feature's full path through the graph.
NA if the state at any of the four time points is NA.}
}
}
\usage{
GRAPH_STATES
}
\description{
\code{repdfr} state assignments (up, down, or null) for each
training-regulated feature (5\% FDR) for each sex at each time point,
which specify node assignments for each differential feature.
Missing values indicate that the \code{repfdr} posterior probabilities
did not meet the cutoff for that feature.
}
\details{
TODO
See \href{https://docs.google.com/document/d/1i5jLyy2K9-N3yMCOvSwnws-5ny77jQPaBtC50TsUZ_g/edit#heading=h.2zazg4neamea}{Supplementary Methods}.
}
\keyword{datasets}
