% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_DA}
\alias{IMMUNO_DA}
\alias{IMMUNO_PLASMA_DA}
\alias{IMMUNO_HIPPOC_DA}
\alias{IMMUNO_CORTEX_DA}
\alias{IMMUNO_ADRNL_DA}
\alias{IMMUNO_TESTES_DA}
\alias{IMMUNO_OVARY_DA}
\alias{IMMUNO_SKMGN_DA}
\alias{IMMUNO_BAT_DA}
\alias{IMMUNO_WATSC_DA}
\alias{IMMUNO_COLON_DA}
\alias{IMMUNO_SMLINT_DA}
\alias{IMMUNO_LIVER_DA}
\alias{IMMUNO_SKMVL_DA}
\alias{IMMUNO_HEART_DA}
\alias{IMMUNO_KIDNEY_DA}
\alias{IMMUNO_SPLEEN_DA}
\alias{IMMUNO_LUNG_DA}
\title{Differential analysis of multiplexed immunoassays}
\format{
A data frame with 17 variables:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link{REPEATED_FEATURES} for details.}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{dataset}}{character, name of LUMINEX panel, e.g., 'rat-mag27plex'}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{adj_p_value}}{double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{covariates}}{character, comma-separated list of adjustment variables or NA}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{selection_fdr}}{double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.}
}

An object of class \code{data.frame} with 368 rows and 17 columns.

An object of class \code{data.frame} with 368 rows and 17 columns.

An object of class \code{data.frame} with 368 rows and 17 columns.

An object of class \code{data.frame} with 184 rows and 17 columns.

An object of class \code{data.frame} with 184 rows and 17 columns.

An object of class \code{data.frame} with 424 rows and 17 columns.

An object of class \code{data.frame} with 424 rows and 17 columns.

An object of class \code{data.frame} with 424 rows and 17 columns.

An object of class \code{data.frame} with 280 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 312 rows and 17 columns.

An object of class \code{data.frame} with 216 rows and 17 columns.

An object of class \code{data.frame} with 208 rows and 17 columns.

An object of class \code{data.frame} with 216 rows and 17 columns.
}
\usage{
IMMUNO_PLASMA_DA

IMMUNO_HIPPOC_DA

IMMUNO_CORTEX_DA

IMMUNO_ADRNL_DA

IMMUNO_TESTES_DA

IMMUNO_OVARY_DA

IMMUNO_SKMGN_DA

IMMUNO_BAT_DA

IMMUNO_WATSC_DA

IMMUNO_COLON_DA

IMMUNO_SMLINT_DA

IMMUNO_LIVER_DA

IMMUNO_SKMVL_DA

IMMUNO_HEART_DA

IMMUNO_KIDNEY_DA

IMMUNO_SPLEEN_DA

IMMUNO_LUNG_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
immunoassay analyte within each sex. One data frame per tissue.
}
\details{
Reproduce our analysis with \code{\link[MotrpacRatTraining6mo:immuno_timewise_da]{MotrpacRatTraining6mo::immuno_timewise_da()}}
and \code{\link[MotrpacRatTraining6mo:immuno_training_da]{MotrpacRatTraining6mo::immuno_training_da()}}.
}
\keyword{datasets}
