% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_NORM_DATA_FLAT}
\alias{IMMUNO_NORM_DATA_FLAT}
\title{Combined immunoassay data used for visualization}
\format{
A data frame with 720 rows and 64 variables:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link{REPEATED_FEATURES} for details.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{dataset}}{character, LUMINEX panel}
}
}
\source{
\code{gs://mawg-data/pass1b-06/immunoassay/data/release/pass1b-06*_mfi-log2-filt-imputed-na-outliers.txt}
}
\usage{
IMMUNO_NORM_DATA_FLAT
}
\description{
Normalized, imputed, and filtered multiplexed immunoassay data used for visualization.
Data are equivalent to the data provided in \link{IMMUNO_NORM_DATA_NESTED}. \link{IMMUNO_NORM_DATA_NESTED} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\details{
Analytes are in rows, and participants IDs (PIDs, i.e., unique animal IDs) are in columns.
}
\keyword{datasets}
