% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMMUNO_NORM_DATA_FLAT}
\alias{IMMUNO_NORM_DATA_FLAT}
\title{Combined immunoassay data used for visualization}
\format{
A data frame with analytes in rows and participant IDs (PIDs) in columns
}
\source{
\code{gs://mawg-data/pass1b-06/immunoassay/data/release/pass1b-06*_mfi-log2-filt-imputed-na-outliers.txt}
}
\usage{
IMMUNO_NORM_DATA_FLAT
}
\description{
Normalized, imputed, and filtered multiplexed immunoassay data used for visualization.
Data are equivalent to the data provided in \link{IMMUNO_NORM_DATA_NESTED}. \link{IMMUNO_NORM_DATA_NESTED} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\keyword{datasets}
