% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_DA}
\alias{METAB_DA}
\alias{METAB_PLASMA_DA}
\alias{METAB_HIPPOC_DA}
\alias{METAB_CORTEX_DA}
\alias{METAB_HYPOTH_DA}
\alias{METAB_SKMGN_DA}
\alias{METAB_SKMVL_DA}
\alias{METAB_HEART_DA}
\alias{METAB_KIDNEY_DA}
\alias{METAB_ADRNL_DA}
\alias{METAB_COLON_DA}
\alias{METAB_SPLEEN_DA}
\alias{METAB_TESTES_DA}
\alias{METAB_OVARY_DA}
\alias{METAB_VENACV_DA}
\alias{METAB_LUNG_DA}
\alias{METAB_SMLINT_DA}
\alias{METAB_LIVER_DA}
\alias{METAB_BAT_DA}
\alias{METAB_WATSC_DA}
\title{Differential analysis of merged metabolomics datasets}
\format{
A data frame with 27 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{dataset}}{character, mame of platform used by the site, e.g. 'metab-u-hilicpos'}
\item{\code{site}}{character, Chemical Analysis Site (CAS) name}
\item{\code{is_targeted}}{logical, is this a targeted platform?}
\item{\code{sex}}{@eval sex()}
\item{\code{comparison_group}}{@eval comparison_group()}
\item{\code{p_value}}{@eval p_value()}
\item{\code{adj_p_value}}{@eval adj_p_value()}
\item{\code{logFC}}{@eval logFC()}
\item{\code{logFC_se}}{@eval logFC_se()}
\item{\code{tscore}}{@eval tscore()}
\item{\code{covariates}}{@eval covariates()}
\item{\code{comparison_average_intensity}}{@eval comparison_average_intensity()}
\item{\code{reference_average_intensity}}{@eval reference_average_intensity()}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite OR the site-given code for unnamed metabolites}
\item{\code{cv}}{double, feature coefficient of variation in the dataset}
\item{\code{metabolite}}{character, name of metabolite as appears in the CAS's data}
\item{\code{control_cv}}{double, feature coefficient of variation in the dataset}
\item{\code{mz}}{double, mass over charge}
\item{\code{rt}}{double, retention time}
\item{\code{neutral_mass}}{double, neutral mass}
\item{\code{selection_fdr}}{@eval selection_fdr()}
}

An object of class \code{data.frame} with 10000 rows and 27 columns.

An object of class \code{data.frame} with 1672 rows and 27 columns.

An object of class \code{data.frame} with 1736 rows and 27 columns.

An object of class \code{data.frame} with 9880 rows and 27 columns.

An object of class \code{data.frame} with 1584 rows and 27 columns.

An object of class \code{data.frame} with 11440 rows and 27 columns.

An object of class \code{data.frame} with 11952 rows and 27 columns.

An object of class \code{data.frame} with 1752 rows and 27 columns.

An object of class \code{data.frame} with 1888 rows and 27 columns.

An object of class \code{data.frame} with 1816 rows and 27 columns.

An object of class \code{data.frame} with 872 rows and 27 columns.

An object of class \code{data.frame} with 912 rows and 27 columns.

An object of class \code{data.frame} with 1278 rows and 27 columns.

An object of class \code{data.frame} with 11736 rows and 27 columns.

An object of class \code{data.frame} with 1800 rows and 27 columns.

An object of class \code{data.frame} with 11896 rows and 27 columns.

An object of class \code{data.frame} with 10402 rows and 27 columns.

An object of class \code{data.frame} with 10288 rows and 27 columns.
}
\usage{
METAB_PLASMA_DA

METAB_HIPPOC_DA

METAB_CORTEX_DA

METAB_HYPOTH_DA

METAB_SKMGN_DA

METAB_SKMVL_DA

METAB_HEART_DA

METAB_KIDNEY_DA

METAB_ADRNL_DA

METAB_COLON_DA

METAB_SPLEEN_DA

METAB_TESTES_DA

METAB_OVARY_DA

METAB_VENACV_DA

METAB_LUNG_DA

METAB_SMLINT_DA

METAB_LIVER_DA

METAB_BAT_DA

METAB_WATSC_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
metabolite within each sex. One data frame per tissue.
}
\keyword{datasets}
