% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_DA_METAREG}
\alias{METAB_DA_METAREG}
\alias{METAB_ADRNL_DA_METAREG}
\alias{METAB_BAT_DA_METAREG}
\alias{METAB_COLON_DA_METAREG}
\alias{METAB_CORTEX_DA_METAREG}
\alias{METAB_HEART_DA_METAREG}
\alias{METAB_HIPPOC_DA_METAREG}
\alias{METAB_HYPOTH_DA_METAREG}
\alias{METAB_KIDNEY_DA_METAREG}
\alias{METAB_LIVER_DA_METAREG}
\alias{METAB_LUNG_DA_METAREG}
\alias{METAB_OVARY_DA_METAREG}
\alias{METAB_PLASMA_DA_METAREG}
\alias{METAB_SKMGN_DA_METAREG}
\alias{METAB_SKMVL_DA_METAREG}
\alias{METAB_SMLINT_DA_METAREG}
\alias{METAB_SPLEEN_DA_METAREG}
\alias{METAB_TESTES_DA_METAREG}
\alias{METAB_VENACV_DA_METAREG}
\alias{METAB_WATSC_DA_METAREG}
\title{Meta-regression of metabolomics differential analysis results}
\format{
A data frame with 30 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{dataset}}{character, mame of platform used by the site, e.g. 'metab-u-hilicpos'}
\item{\code{site}}{character, Chemical Analysis Site (CAS) name}
\item{\code{is_targeted}}{logical, is this a targeted platform?}
\item{\code{sex}}{@eval sex()}
\item{\code{comparison_group}}{@eval comparison_group()}
\item{\code{p_value}}{@eval p_value()}
\item{\code{adj_p_value}}{@eval adj_p_value()}
\item{\code{logFC}}{@eval logFC()}
\item{\code{logFC_se}}{@eval logFC_se()}
\item{\code{tscore}}{@eval tscore()}
\item{\code{zscore}}{@eval zscore()}
\item{\code{covariates}}{@eval covariates()}
\item{\code{comparison_average_intensity}}{@eval comparison_average_intensity()}
\item{\code{reference_average_intensity}}{@eval reference_average_intensity()}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite OR the site-given code for unnamed metabolites}
\item{\code{cv}}{double, feature coefficient of variation in the dataset}
\item{\code{metabolite}}{character, name of metabolite as appears in the CAS's data}
\item{\code{control_cv}}{double, feature coefficient of variation in the dataset}
\item{\code{mz}}{double, mass over charge}
\item{\code{rt}}{double, retention time}
\item{\code{neutral_mass}}{double, neutral mass}
\item{\code{meta_reg_het_p}}{@eval meta_reg_het_p()}
\item{\code{meta_reg_pvalue}}{@eval meta_reg_pvalue()}
\item{\code{selection_fdr}}{@eval selection_fdr()}
}

An object of class \code{data.frame} with 9086 rows and 30 columns.

An object of class \code{data.frame} with 1888 rows and 30 columns.

An object of class \code{data.frame} with 1672 rows and 30 columns.

An object of class \code{data.frame} with 9872 rows and 30 columns.

An object of class \code{data.frame} with 8920 rows and 30 columns.

An object of class \code{data.frame} with 1736 rows and 30 columns.

An object of class \code{data.frame} with 10376 rows and 30 columns.

An object of class \code{data.frame} with 10264 rows and 30 columns.

An object of class \code{data.frame} with 10208 rows and 30 columns.

An object of class \code{data.frame} with 912 rows and 30 columns.

An object of class \code{data.frame} with 8696 rows and 30 columns.

An object of class \code{data.frame} with 8472 rows and 30 columns.

An object of class \code{data.frame} with 1584 rows and 30 columns.

An object of class \code{data.frame} with 1800 rows and 30 columns.

An object of class \code{data.frame} with 1816 rows and 30 columns.

An object of class \code{data.frame} with 872 rows and 30 columns.

An object of class \code{data.frame} with 1278 rows and 30 columns.

An object of class \code{data.frame} with 8808 rows and 30 columns.
}
\usage{
METAB_ADRNL_DA_METAREG

METAB_BAT_DA_METAREG

METAB_COLON_DA_METAREG

METAB_CORTEX_DA_METAREG

METAB_HEART_DA_METAREG

METAB_HIPPOC_DA_METAREG

METAB_HYPOTH_DA_METAREG

METAB_KIDNEY_DA_METAREG

METAB_LIVER_DA_METAREG

METAB_LUNG_DA_METAREG

METAB_OVARY_DA_METAREG

METAB_PLASMA_DA_METAREG

METAB_SKMGN_DA_METAREG

METAB_SKMVL_DA_METAREG

METAB_SMLINT_DA_METAREG

METAB_SPLEEN_DA_METAREG

METAB_TESTES_DA_METAREG

METAB_VENACV_DA_METAREG

METAB_WATSC_DA_METAREG
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
metabolite within each sex. One data frame per tissue.

For metabolites measured on more than one
platform, the consensus result may be provided. For metabolites measured
on a single platform, these results are identical to \link{METAB_DA}.
}
\details{
DETAILS
}
\keyword{datasets}
