% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_NORM_DATA_FLAT}
\alias{METAB_NORM_DATA_FLAT}
\title{Combined metabolomics data used for visualization}
\format{
A data frame with 59 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{assay}}{@eval assay()}
\item{\code{dataset}}{@eval dataset_metab()}
\item{\code{10023259}}{@eval 10023259()}
\item{\code{10024735}}{@eval 10024735()}
\item{\code{10025626}}{@eval 10025626()}
\item{\code{10025707}}{@eval 10025707()}
\item{\code{10025979}}{@eval 10025979()}
\item{\code{10026193}}{@eval 10026193()}
\item{\code{10026355}}{@eval 10026355()}
\item{\code{10026789}}{@eval 10026789()}
\item{\code{10027327}}{@eval 10027327()}
\item{\code{10027599}}{@eval 10027599()}
\item{\code{10043527}}{@eval 10043527()}
\item{\code{10043950}}{@eval 10043950()}
\item{\code{10044256}}{@eval 10044256()}
\item{\code{10044337}}{@eval 10044337()}
\item{\code{10044760}}{@eval 10044760()}
\item{\code{10044841}}{@eval 10044841()}
\item{\code{10044922}}{@eval 10044922()}
\item{\code{10045228}}{@eval 10045228()}
\item{\code{10045309}}{@eval 10045309()}
\item{\code{10046119}}{@eval 10046119()}
\item{\code{10046380}}{@eval 10046380()}
\item{\code{10046461}}{@eval 10046461()}
\item{\code{10095241}}{@eval 10095241()}
\item{\code{10106383}}{@eval 10106383()}
\item{\code{10139044}}{@eval 10139044()}
\item{\code{10185585}}{@eval 10185585()}
\item{\code{10204989}}{@eval 10204989()}
\item{\code{10220917}}{@eval 10220917()}
\item{\code{10306471}}{@eval 10306471()}
\item{\code{10314733}}{@eval 10314733()}
\item{\code{10335064}}{@eval 10335064()}
\item{\code{10337199}}{@eval 10337199()}
\item{\code{10381414}}{@eval 10381414()}
\item{\code{10411690}}{@eval 10411690()}
\item{\code{10417702}}{@eval 10417702()}
\item{\code{10422463}}{@eval 10422463()}
\item{\code{10424911}}{@eval 10424911()}
\item{\code{10486003}}{@eval 10486003()}
\item{\code{10502300}}{@eval 10502300()}
\item{\code{10503110}}{@eval 10503110()}
\item{\code{10561072}}{@eval 10561072()}
\item{\code{10587543}}{@eval 10587543()}
\item{\code{10641858}}{@eval 10641858()}
\item{\code{10672656}}{@eval 10672656()}
\item{\code{10700102}}{@eval 10700102()}
\item{\code{10729518}}{@eval 10729518()}
\item{\code{10734945}}{@eval 10734945()}
\item{\code{10761160}}{@eval 10761160()}
\item{\code{10833331}}{@eval 10833331()}
\item{\code{10871349}}{@eval 10871349()}
\item{\code{10901434}}{@eval 10901434()}
\item{\code{10950486}}{@eval 10950486()}
\item{\code{10953744}}{@eval 10953744()}
\item{\code{10971645}}{@eval 10971645()}
}
}
\usage{
METAB_NORM_DATA_FLAT
}
\description{
Combined sample-level metabolomics data used for visualization.
Data are equivalent to the data provided in \link{METAB_SAMPLE_DATA_NESTED}. \link{METAB_SAMPLE_DATA_NESTED} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\details{
TODO
}
\keyword{datasets}
