% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_NORM_DATA_FLAT}
\alias{METAB_NORM_DATA_FLAT}
\title{Combined metabolomics data used for visualization}
\format{
A data frame with metabolites in rows and participant IDs (PIDs) in columns
}
\usage{
METAB_NORM_DATA_FLAT
}
\description{
Combined sample-level metabolomics data used for visualization.
Data are equivalent to the data provided in \link{METAB_NORM_DATA_NESTED}. \link{METAB_NORM_DATA_NESTED} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\keyword{datasets}
