% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_NORM_DATA_FLAT}
\alias{METAB_NORM_DATA_FLAT}
\title{Combined metabolomics data used for visualization}
\format{
A data frame with 59 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{assay}}{@eval assay()}
\item{\code{dataset}}{@eval dataset_metab()}
}
}
\usage{
METAB_NORM_DATA_FLAT
}
\description{
Combined sample-level metabolomics data used for visualization.
Data are equivalent to the data provided in \link{METAB_SAMPLE_DATA_NESTED}. \link{METAB_SAMPLE_DATA_NESTED} is compatible with the
differential analysis functions while this format is compatible with visualization functions.
}
\details{
TODO
}
\keyword{datasets}
