% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_SAMPLE_DATA}
\alias{METAB_SAMPLE_DATA}
\title{Metabolomics normalized sample-level data}
\format{
A data frame with 25477 rows and 55 columns
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/metabolomics-named-merged/normalized-data}
}
\usage{
METAB_SAMPLE_DATA
}
\description{
Normalized sample-level metabolomics data

For targeted datasets, the file matching "convert2na-log2-featurectr-featurefilt-knn" was used if it existed; "named-convert2na-log2.txt" was used otherwise.

For untargeted datasets, the decision table ("gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/metabolomics-named-merged/stats-tests/pass1b-06_sample-ctr-decision-table-kw-summary.txt")
was used to determine which normalized file to use for each dataset. If the data were supposed to be sample-centered,
the file matching "named-featurefilt-knn-samplefilt-log2-featurestd-samplectr.txt" was used. Otherwise, "named-featurefilt-knn-samplefilt-log2-featurestd.txt" was used.
}
\details{
One row per feature, a \code{feature} column, and one column per PID (participant ID).
\code{feature} is in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}.
}
\keyword{datasets}
