% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{METHYL_FEATURE_ANNOT}
\alias{METHYL_FEATURE_ANNOT}
\title{RRBS feature annotation}
\format{
A data frame with 7585076 rows and 12 variables:
\describe{
\item{\code{Chr}}{integer, chromosome}
\item{\code{Locus}}{character, base pair range of feature}
\item{\code{EntrezID}}{character, Entrez ID of closest gene}
\item{\code{Symbol}}{character, gene symbol of closest gene}
\item{\code{Strand}}{character, strand}
\item{\code{Width}}{integer, width of feature}
\item{\code{NumSites}}{integer, TODO}
\item{\code{Sites}}{character, TODO}
\item{\code{LocStart}}{integer, feature start in base pairs}
\item{\code{LocEnd}}{integer, feature end in base pairs}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
}
}
\description{
RRBS feature annotation
}
\details{
TODO
}
