% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OUTLIERS}
\alias{OUTLIERS}
\title{Sample outliers}
\format{
A data frame with 27 rows and 9 variables:
\describe{
\item{\code{viallabel}}{character, sample identifier}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{platform}}{character, LUMINEX panel if \code{assay} is IMMUNO}
\item{\code{pid}}{integer, participant ID, one per animal}
\item{\code{group}}{character, combination of sex and training time point that the sample belongs to, e.g., "female_1w"}
\item{\code{reason}}{character, reason(s) the sample was called an outlier}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1DetAMovcmMJqulEA41yBLd4I9Q1XSHPH1U8U2aVMShg/edit#gid=2057863805}
}
\usage{
OUTLIERS
}
\description{
Outliers excluded during differential analysis
}
\details{
If the sample was an outlier in principal component (PC) space, \code{reason}
lists the PC(s) in which it was an outlier. See ome-specific details of outlier calling
in the supplementary methods of the \href{https://github.com/MoTrPAC/MotrpacRatTraining6moData#citing-motrpac-data}{manuscript}.
}
\keyword{datasets}
