% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHENO}
\alias{PHENO}
\title{Phenotypic data}
\format{
A data frame with 5955 rows and 509 variables:
\describe{
\item{\code{pid}}{integer, unique, randomly generated 8 digit numeric identifier used in linkage to phenotypic data}
\item{\code{bid}}{integer, unique, randomly generated 8 digit numeric identifier used in linkage to phenotypic data}
\item{\code{labelid}}{double, unique 11 digit identifier for specimen label ID, originating at the collection site,
that provides a link to specimen processing and used for shipments to the biorepository (same as
\code{vialLabel} only in instances where aliquots were not further processed at the biorepository)}
\item{\code{viallabel}}{character, unique 11 digit numeric identifier of a sample vial made up of \code{bid},
\code{key.sacrificetime}, \code{specimen.processing.sampletypedescription}, and aliquot number}
\item{\code{key.protocol}}{character, protocol in which the procedures for sample collection were outlined}
\item{\code{key.agegroup}}{character, age group for the animal}
\item{\code{key.d_arrive}}{character, date animal arrived at site (same as \code{registration.d_arrive})}
\item{\code{key.d_sacrifice}}{character, date animal was sacrificed (same as \code{specimen.collection.d_visit})}
\item{\code{key.intervention}}{character, exercise group defined by those who complete exercise training}
\item{\code{key.sacrificetime}}{character, weeks since the beginning of the intervention}
\item{\code{key.anirandgroup}}{character, description of the randomization group}
\item{\code{key.siteid}}{character, site where the samples were collected}
\item{\code{registration.d_visit}}{character, date registration was completed}
\item{\code{registration.days_visit}}{integer, day registration was completed relative to \code{registration.d_arrive} (count)}
\item{\code{registration.staffid}}{integer, registration staff ID}
\item{\code{registration.siteid}}{character, registration site}
\item{\code{registration.formname}}{character, registration form name}
\item{\code{registration.versionnbr}}{integer, registration form version number}
\item{\code{registration.ratid}}{character, site rat ID}
\item{\code{registration.d_arrive}}{character, date the rat arrived at the site}
\item{\code{registration.days_arrive}}{integer, day of arrival relative to \code{registration.d_arrive} (count)}
\item{\code{registration.batchnumber}}{integer, registration batch number}
\item{\code{registration.d_reverselight}}{character, date the rat was housed in reverse light}
\item{\code{registration.days_reverselight}}{integer, day the rat was housed in reverse light relative to \code{registration.d_arrive} (count)}
\item{\code{registration.d_birth}}{character, rat date of birth (+/- two weeks)}
\item{\code{registration.days_birth}}{integer, rat day of birth relative to \code{registration.d_arrive} (count)}
\item{\code{registration.sex}}{character, sex, one of "male", "female"}
\item{\code{registration.weight}}{double, rat's body weight}
\item{\code{registration.cagenumber}}{integer, cage number in which the rat was housed}
\item{\code{familiarization.d_visit}}{character, date familiarization was completed}
\item{\code{familiarization.days_visit}}{integer, day familiarization was completed relative to \code{registration.d_arrive} (count)}
\item{\code{familiarization.staffid}}{integer, familiarization staff ID}
\item{\code{familiarization.siteid}}{character, familiarization site}
\item{\code{familiarization.formname}}{character, familiarization form name}
\item{\code{familiarization.versionnbr}}{integer, familiarization form version number}
\item{\code{familiarization.d_treadmillbegin}}{character, date the rat began treadmill familiarization}
\item{\code{familiarization.days_treadmillbegin}}{integer, day the rat began treadmill familiarization relative to \code{registration.d_arrive} (count)}
\item{\code{familiarization.d_treadmillcomplete}}{character, date the rat completed treadmill familiarization}
\item{\code{familiarization.days_treadmillcomplete}}{integer, day the rat completed treadmill familiarization relative to \code{registration.d_arrive} (count)}
\item{\code{familiarization.activity_score}}{double, rat's activity score}
\item{\code{familiarization.compliant}}{character, was this rat compliant?}
\item{\code{familiarization.weight}}{double, rat's body weight}
\item{\code{training.d_visit}}{character, date intervention started}
\item{\code{training.days_visit}}{integer, day intervention started relative to \code{registration.d_arrive}}
\item{\code{training.staffid}}{integer, training staff ID}
\item{\code{training.siteid}}{character, training site}
\item{\code{training.formname}}{character, training form name}
\item{\code{training.versionnbr}}{integer, training form version number}
\item{\code{training.day*date}}{character, date of given training day}
\item{\code{training.day*_days}}{integer, day of given training day relative to \code{registration.d_arrive} (count)}
\item{\code{training.day*time}}{character, time of given training day}
\item{\code{training.day*_treadmillspeed}}{double, treadmill speed on given training day}
\item{\code{training.day*_treadmillincline}}{double, treadmill incline on given training day}
\item{\code{training.day*_timeontreadmill}}{integer, time on treadmill on given training day}
\item{\code{training.day*_weight}}{double, weight on given training day}
\item{\code{training.day*_posttrainlact}}{double, post-training lactate on given training day}
\item{\code{training.day*_score}}{integer, score on given training day}
\item{\code{training.day*_comments}}{character, comments on given training day}
\item{\code{vo2.max.test.d_visit_1}}{character, date first VO2 max test was completed}
\item{\code{vo2.max.test.d_visit_2}}{character, date second VO2 max test was completed}
\item{\code{vo2.max.test.days_visit_1}}{integer, day first VO2 max test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.days_visit_2}}{integer, day second VO2 max test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.staffid_1}}{integer, first VO2 max test staff ID}
\item{\code{vo2.max.test.staffid_2}}{integer, second VO2 max test staff ID}
\item{\code{vo2.max.test.siteid_1}}{character, first VO2 max test site}
\item{\code{vo2.max.test.siteid_2}}{character, second VO2 max test site}
\item{\code{vo2.max.test.visitcode_1}}{character, visit code for first VO2 max test}
\item{\code{vo2.max.test.visitcode_2}}{character, visit code for second VO2 max test}
\item{\code{vo2.max.test.formname_1}}{character, form name for first VO2 max test}
\item{\code{vo2.max.test.formname_2}}{character, form name for second VO2 max test}
\item{\code{vo2.max.test.versionnbr_1}}{integer, form version number for first VO2 max test}
\item{\code{vo2.max.test.versionnbr_2}}{integer, form version number for second VO2 max test}
\item{\code{vo2.max.test.d_vo2_1}}{character, date of first VO2 max test}
\item{\code{vo2.max.test.d_vo2_2}}{character, date of second VO2 max test}
\item{\code{vo2.max.test.days_vo2_1}}{integer, date of first VO2 max test relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.days_vo2_2}}{integer, date of second VO2 max test relative to \code{registration.d_arrive} (count)}
\item{\code{vo2.max.test.t_vo2_1}}{character, start time of first VO2 max test?}
\item{\code{vo2.max.test.t_vo2_2}}{character, start time of second VO2 max test?}
\item{\code{vo2.max.test.blactate_begin_1}}{double, blood lactate at the beginning of exercise for first VO2 max test}
\item{\code{vo2.max.test.blactate_begin_2}}{double, blood lactate at the beginning of exercise for second VO2 max test}
\item{\code{vo2.max.test.vo2_max_1}}{double, VO2 at max for first VO2 max test}
\item{\code{vo2.max.test.vo2_max_2}}{double, VO2 at max for second VO2 max test}
\item{\code{vo2.max.test.vco2_max_1}}{double, VCO2 at max for first VO2 max test}
\item{\code{vo2.max.test.vco2_max_2}}{double, VCO2 at max for second VO2 max test}
\item{\code{vo2.max.test.rer_max_1}}{double, RER at max for first VO2 max test}
\item{\code{vo2.max.test.rer_max_2}}{double, RER at max for second VO2 max test}
\item{\code{vo2.max.test.speed_max_1}}{double, max speed of running for first VO2 max test}
\item{\code{vo2.max.test.speed_max_2}}{double, max speed of running for second VO2 max test}
\item{\code{vo2.max.test.t_complete_1}}{character, time of completion for first VO2 max test}
\item{\code{vo2.max.test.t_complete_2}}{character, time of completion for second VO2 max test}
\item{\code{vo2.max.test.blactate_end_1}}{double, blood lactate at end of exercise for first VO2 max test}
\item{\code{vo2.max.test.blactate_end_2}}{double, blood lactate at end of exercise for second VO2 max test}
\item{\code{vo2.max.test.vo2_comments_1}}{character, comments about first VO2 max test}
\item{\code{vo2.max.test.vo2_comments_2}}{character, comments about second VO2 max test}
\item{\code{nmr.testing.d_visit_1}}{character, date first NMR test was completed}
\item{\code{nmr.testing.d_visit_2}}{character, date second NMR test was completed}
\item{\code{nmr.testing.days_visit_1}}{integer, day first NMR test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.days_visit_2}}{integer, day second NMR test was completed relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.staffid_1}}{integer, first NMR test staff ID}
\item{\code{nmr.testing.staffid_2}}{integer, second NMR test staff ID}
\item{\code{nmr.testing.siteid_1}}{character, first NMR test site}
\item{\code{nmr.testing.siteid_2}}{character, second NMR test site}
\item{\code{nmr.testing.visitcode_1}}{character, visit code for first NMR test}
\item{\code{nmr.testing.visitcode_2}}{character, visit code for second NMR test}
\item{\code{nmr.testing.formname_1}}{character, form name for first NMR test}
\item{\code{nmr.testing.formname_2}}{character, form name for second NMR test}
\item{\code{nmr.testing.versionnbr_1}}{integer, form version number for first NMR test}
\item{\code{nmr.testing.versionnbr_2}}{integer, form version number for second NMR test}
\item{\code{nmr.testing.d_nmr_1}}{character, date data was collected for the first NMR test}
\item{\code{nmr.testing.d_nmr_2}}{character, date data was collected for the second NMR test}
\item{\code{nmr.testing.days_nmr_1}}{integer, day data was collected for the first NMR test relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.days_nmr_2}}{integer, day data was collected for the second NMR test relative to \code{registration.d_arrive} (count)}
\item{\code{nmr.testing.nmr_weight_1}}{double, rat's body weight for first NMR test}
\item{\code{nmr.testing.nmr_weight_2}}{double, rat's body weight for second NMR test}
\item{\code{nmr.testing.nmr_fat_1}}{double, rat's percent body fat for first NMR test}
\item{\code{nmr.testing.nmr_fat_2}}{double, rat's percent body fat for second NMR test}
\item{\code{nmr.testing.nmr_lean_1}}{double, rat's percent lean for first NMR test}
\item{\code{nmr.testing.nmr_lean_2}}{double, rat's percent lean for second NMR test}
\item{\code{nmr.testing.nmr_fluid_1}}{double, rat's percent fluid for first NMR test}
\item{\code{nmr.testing.nmr_fluid_2}}{double, rat's percent fluid for second NMR test}
\item{\code{nmr.testing.nmr_comments_2}}{character, comments about second NMR test}
\item{\code{terminal.weight.bw}}{double, terminal body weight (g)}
\item{\code{terminal.weight.lg}}{integer, terminal lateral gastrocnemius weight (mg)}
\item{\code{terminal.weight.mg}}{integer, terminal medial gastrocnemius weight (mg)}
\item{\code{terminal.weight.pl}}{integer, terminal plantaris weight (mg)}
\item{\code{terminal.weight.sol}}{integer, terminal soleus weight (mg)}
\item{\code{specimen.collection.d_visit}}{character, date of specimen collection}
\item{\code{specimen.collection.days_visit}}{integer, day of specimen collection relative to \code{registration.d_arrive} (count)}
\item{\code{specimen.collection.staffid}}{integer, specimen collection staff ID}
\item{\code{specimen.collection.siteid}}{character, specimen collection site}
\item{\code{specimen.collection.visitcode}}{character, specimen collection site visit code}
\item{\code{specimen.collection.formname}}{character, specimen collection form name}
\item{\code{specimen.collection.versionnbr}}{integer, specimen collection form version number}
\item{\code{specimen.collection.anesthesiaid}}{integer, anesthesia administrator ID}
\item{\code{specimen.collection.t_anesthesia}}{character, anesthesia administration time}
\item{\code{specimen.collection.bloodtype}}{character, blood type}
\item{\code{specimen.collection.bloodtube}}{character, blood tube}
\item{\code{specimen.collection.bloodcomplete}}{character, blood collection complete}
\item{\code{specimen.collection.bloodtechid}}{character, blood collection technician ID}
\item{\code{specimen.collection.t_bloodstart}}{character, blood collection start time}
\item{\code{specimen.collection.t_bloodstop}}{character, blood collection stop time}
\item{\code{specimen.collection.t_edtafill}}{character, EDTA fill time}
\item{\code{specimen.collection.uterustype}}{character, uterus type}
\item{\code{specimen.collection.uterustechid}}{character, uterus technician ID}
\item{\code{specimen.collection.uteruscomplete}}{character, uterus complete}
\item{\code{specimen.collection.t_death}}{character, time of death}
\item{\code{specimen.collection.deathtype}}{character, type of death}
\item{\code{specimen.processing.versionnbr}}{integer, specimen processing form version number}
\item{\code{specimen.processing.formname}}{character, specimen processing form name}
\item{\code{specimen.processing.visitcode}}{character, specimen processing visit code}
\item{\code{specimen.processing.siteid}}{character, specimen processing site}
\item{\code{specimen.processing.sampletypedescription}}{character, specimen tissue type}
\item{\code{specimen.processing.samplenumber}}{integer, specimen sample number}
\item{\code{specimen.processing.timepoint}}{integer, specimen time point}
\item{\code{specimen.processing.aliquotdescription}}{character, specimen aliquot tube}
\item{\code{specimen.processing.volume}}{character, specimen aliquot volume}
\item{\code{specimen.processing.hemolyzed}}{character, specimen hemolyzed?}
\item{\code{specimen.processing.t_collection}}{character, specimen collection time}
\item{\code{specimen.processing.t_edtaspin}}{character, EDTA spin time}
\item{\code{specimen.processing.t_freeze}}{character, specimen freeze time}
\item{\code{specimen.processing.techid}}{character, specimen processing tech ID}
\item{\code{calculated.variables.wgt_gain_after_train}}{double, weight gain after training:
\code{max(nmr.testing.nmr_weight_2, nmr.testing.nmr_weight_1) - registration.weight}}
\item{\code{calculated.variables.pct_body_fat_change}}{double, percent body fat change after training:
\code{nmr.testing.nmr_fat_2 - nmr.testing.nmr_fat_1}}
\item{\code{calculated.variables.pct_body_lean_change}}{double, percent body lean change after training:
\code{nmr.testing.nmr_lean_2 - nmr.testing.nmr_lean_1}}
\item{\code{calculated.variables.pct_body_fluid_change}}{double, percent body fluid change after training:
\code{nmr.testing.nmr_fluid_2 - nmr.testing.nmr_fluid_1}}
\item{\code{calculated.variables.lactate_change_dueto_train}}{double, lactate change due to training:
\code{vo2.max.test.blactate_end_2 - vo2.max.test.blactate_begin_1}}
\item{\code{calculated.variables.vo2_max_change}}{double, change in VO2 max after training:
\code{vo2.max.test.vo2_max_2 - vo2.max.test.vo2_max_1}}
\item{\code{calculated.variables.coll_time_train}}{integer, sample collection time after training (seconds):
specimen collection date and time - last time trained}
\item{\code{calculated.variables.deathtime_after_train}}{integer, time of death after training (seconds):
death date and time - last time trained}
\item{\code{calculated.variables.frozetime_after_train}}{integer, time of sample freezing after training (seconds):
specimen collection date and freeze time - last time trained}
\item{\code{biclabeldata.shiptositeid}}{character, Chemical Analysis Site (CAS), i.e. the omics site)}
\item{\code{sacrificetime}}{character, sacrifice time point in weeks, one of "1w", "2w", "4w", "8w"}
\item{\code{intervention}}{character, intervention, one of "training", "control"}
\item{\code{group}}{character, intervention group, one of "1w", "2w", "4w", "8w", "control"}
\item{\code{sex}}{character, sex, one of "male", "female"}
\item{\code{time_to_freeze}}{integer, sample time to freeze, \code{calculated.variables.frozetime_after_train - calculated.variables.deathtime_after_train}}
}
}
\usage{
PHENO
}
\description{
Phenotypic data for samples from the MoTrPAC endurance exercise
training study in 6-month-old rats. One row per sample (\code{viallabel}).
}
\keyword{datasets}
