% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_NORM_DATA}
\alias{PROT_NORM_DATA}
\alias{PROT_CORTEX_NORM_DATA}
\alias{PROT_SKMGN_NORM_DATA}
\alias{PROT_HEART_NORM_DATA}
\alias{PROT_KIDNEY_NORM_DATA}
\alias{PROT_LUNG_NORM_DATA}
\alias{PROT_LIVER_NORM_DATA}
\alias{PROT_WATSC_NORM_DATA}
\title{Normalized protein expression data}
\format{
A data frame with protein IDs in rows (\code{feature_ID}) and samples in columns (\code{viallabel})

An object of class \code{data.frame} with 5999 rows and 58 columns.

An object of class \code{data.frame} with 9184 rows and 61 columns.

An object of class \code{data.frame} with 9852 rows and 61 columns.

An object of class \code{data.frame} with 10877 rows and 61 columns.

An object of class \code{data.frame} with 9791 rows and 60 columns.

An object of class \code{data.frame} with 9964 rows and 61 columns.
}
\source{
\code{gs://motrpac-data-freeze-pass/pass1b-06/v1.1/analysis/proteomics-untargeted/prot-pr/normalized-data/*med-mad-normalized-logratio.txt}
}
\usage{
PROT_CORTEX_NORM_DATA

PROT_SKMGN_NORM_DATA

PROT_HEART_NORM_DATA

PROT_KIDNEY_NORM_DATA

PROT_LUNG_NORM_DATA

PROT_LIVER_NORM_DATA

PROT_WATSC_NORM_DATA
}
\description{
Median-MAD normalized protein expression (PROT) data used for visualization and differential analysis
}
\details{
Log2 TMT ratios to the common reference were used as quantitative values for all proteomics features
(proteins, phosphosites, acetylsite, and ubiquitylsites). Proteomics features not fully quantified in at least
two plexes within a tissue and non-rat contaminants were removed. Log2 TMT ratios were sample-normalized by
median-centering and mean absolute deviation scaling. Plex batch effects were removed using linear models
implemented by the \code{\link[limma:removeBatchEffect]{limma::removeBatchEffect()}}.
}
\keyword{datasets}
