% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROTEOME_DA}
\alias{PROTEOME_DA}
\alias{PROT_CORTEX_DA}
\alias{PROT_SKMGN_DA}
\alias{PROT_HEART_DA}
\alias{PROT_KIDNEY_DA}
\alias{PROT_LUNG_DA}
\alias{PROT_LIVER_DA}
\alias{PROT_WATSC_DA}
\alias{PHOSPHO_CORTEX_DA}
\alias{PHOSPHO_SKMGN_DA}
\alias{PHOSPHO_HEART_DA}
\alias{PHOSPHO_KIDNEY_DA}
\alias{PHOSPHO_LUNG_DA}
\alias{PHOSPHO_LIVER_DA}
\alias{PHOSPHO_WATSC_DA}
\alias{ACETYL_HEART_DA}
\alias{ACETYL_LIVER_DA}
\alias{UBIQ_HEART_DA}
\alias{UBIQ_LIVER_DA}
\title{Differential analysis of proteomics datasets}
\format{
A data frame 18 variables:
\describe{
\item{\code{feature}}{character, unique feature identifier in the format '\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};feature_ID' only for training-regulated features at 5\% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See \link{REPEATED_FEATURES} for details.}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{assay_code}}{character, assay code used in data release. See \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{tissue_code}}{character, tissue code used in data release. See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code}.}
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{sex}}{character, one of 'male' or 'female'}
\item{\code{comparison_group}}{character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'}
\item{\code{p_value}}{double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals}
\item{\code{adj_p_value}}{double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.}
\item{\code{logFC}}{double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{covariates}}{character, comma-separated list of adjustment variables or NA}
\item{\code{numNAs}}{integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in 'comparison_group'}
\item{\code{reference_average_intensity}}{double, average intensity among the replicates in the group of sex-matched sedentary control animals}
\item{\code{selection_fdr}}{double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.}
}

An object of class \code{data.frame} with 47992 rows and 18 columns.

An object of class \code{data.frame} with 73472 rows and 18 columns.

An object of class \code{data.frame} with 78816 rows and 18 columns.

An object of class \code{data.frame} with 87016 rows and 18 columns.

An object of class \code{data.frame} with 78328 rows and 18 columns.

An object of class \code{data.frame} with 79712 rows and 18 columns.

An object of class \code{data.frame} with 372016 rows and 18 columns.

An object of class \code{data.frame} with 237208 rows and 18 columns.

An object of class \code{data.frame} with 321664 rows and 18 columns.

An object of class \code{data.frame} with 241152 rows and 18 columns.

An object of class \code{data.frame} with 381288 rows and 18 columns.

An object of class \code{data.frame} with 348536 rows and 18 columns.

An object of class \code{data.frame} with 242432 rows and 18 columns.

An object of class \code{data.frame} with 41704 rows and 18 columns.

An object of class \code{data.frame} with 78000 rows and 18 columns.

An object of class \code{data.frame} with 56624 rows and 18 columns.

An object of class \code{data.frame} with 74752 rows and 18 columns.
}
\usage{
PROT_CORTEX_DA

PROT_SKMGN_DA

PROT_HEART_DA

PROT_KIDNEY_DA

PROT_LUNG_DA

PROT_LIVER_DA

PROT_WATSC_DA

PHOSPHO_CORTEX_DA

PHOSPHO_SKMGN_DA

PHOSPHO_HEART_DA

PHOSPHO_KIDNEY_DA

PHOSPHO_LUNG_DA

PHOSPHO_LIVER_DA

PHOSPHO_WATSC_DA

ACETYL_HEART_DA

ACETYL_LIVER_DA

UBIQ_HEART_DA

UBIQ_LIVER_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
proteomics feature (protein, phosphosite, acetylsite, ubiquitlsite)
within each sex. One data frame per data type per tissue.
}
\details{
Reproduce our analysis with \code{\link[MotrpacRatTraining6mo:proteomics_timewise_da]{MotrpacRatTraining6mo::proteomics_timewise_da()}}
and \code{\link[MotrpacRatTraining6mo:proteomics_training_da]{MotrpacRatTraining6mo::proteomics_training_da()}}.
}
\keyword{datasets}
