% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROTEOME_DA}
\alias{PROTEOME_DA}
\alias{PROT_CORTEX_DA}
\alias{PROT_SKMGN_DA}
\alias{PROT_HEART_DA}
\alias{PROT_KIDNEY_DA}
\alias{PROT_LUNG_DA}
\alias{PROT_LIVER_DA}
\alias{PROT_WATSC_DA}
\alias{PHOSPHO_CORTEX_DA}
\alias{PHOSPHO_SKMGN_DA}
\alias{PHOSPHO_HEART_DA}
\alias{PHOSPHO_KIDNEY_DA}
\alias{PHOSPHO_LUNG_DA}
\alias{PHOSPHO_LIVER_DA}
\alias{PHOSPHO_WATSC_DA}
\alias{ACETYL_HEART_DA}
\alias{ACETYL_LIVER_DA}
\alias{UBIQ_HEART_DA}
\alias{UBIQ_LIVER_DA}
\title{Differential analysis of proteomics datasets}
\format{
A data frame with >40000 rows and 16 variables:
\describe{
\item{\code{feature_ID}}{character, analyte identifier}
\item{\code{tissue}}{character, MoTrPAC tissue release code}
\item{\code{assay}}{character, MoTrPAC assay or "ome" code. "prot-pr" for
global proteomics; "prot-ph" for phosphoproteomics; "prot-ac" for
acetylomics; "prot-ub" for ubiquitylomics.}
\item{\code{sex}}{character, one of "male" or "female"}
\item{\code{logFC}}{double, log fold-change where the numerator is \code{comparison_group}
and the denominator is the group of sex-matched sedentary control animals}
\item{\code{logFC_se}}{double, standard error of the log fold-change}
\item{\code{tscore}}{double, t statistic}
\item{\code{covariates}}{character, comma-separated list of covariates or NA}
\item{\code{comparison_group}}{character, time point compared to the sex-matched
sedentary control animals, one of "1w", "2w", "4w", "8w"}
\item{\code{p_value}}{double, unadjusted p-value for the difference between
\code{comparison_group} and the group of sex-matched sedentary control animals}
\item{\code{comparison_average_intensity}}{double, average intensity among the replicates in \code{comparison_group}}
\item{\code{reference_average_intensity}}{double, average intensity among the
replicates in the group of sex-matched sedentary control animals}
\item{\code{numNAs}}{integer, number of missing values in the current
\code{comparison_group} and sex-matched sedentary control animals}
\item{\code{adj_p_value}}{double, adjusted p-value from \code{p_value} column.
P-values are BY-adjusted across all datasets within a given \code{assay}.}
\item{\code{tissue_abbreviation}}{character, MoTrPAC tissue abbreviation used in manuscripts}
\item{\code{selection_fdr}}{double, adjusted training p-value used to
select training-regulated analytes. P-values are IHW-adjusted across all
datasets within a given \code{assay} with \code{tissue} as a covariate.}
}

An object of class \code{data.frame} with 47992 rows and 16 columns.

An object of class \code{data.frame} with 73472 rows and 16 columns.

An object of class \code{data.frame} with 78816 rows and 16 columns.

An object of class \code{data.frame} with 87016 rows and 16 columns.

An object of class \code{data.frame} with 78328 rows and 16 columns.

An object of class \code{data.frame} with 79712 rows and 16 columns.

An object of class \code{data.frame} with 372016 rows and 16 columns.

An object of class \code{data.frame} with 237208 rows and 16 columns.

An object of class \code{data.frame} with 321664 rows and 16 columns.

An object of class \code{data.frame} with 241152 rows and 16 columns.

An object of class \code{data.frame} with 381288 rows and 16 columns.

An object of class \code{data.frame} with 348536 rows and 16 columns.

An object of class \code{data.frame} with 242432 rows and 16 columns.

An object of class \code{data.frame} with 41704 rows and 16 columns.

An object of class \code{data.frame} with 78000 rows and 16 columns.

An object of class \code{data.frame} with 56624 rows and 16 columns.

An object of class \code{data.frame} with 74752 rows and 16 columns.
}
\usage{
PROT_CORTEX_DA

PROT_SKMGN_DA

PROT_HEART_DA

PROT_KIDNEY_DA

PROT_LUNG_DA

PROT_LIVER_DA

PROT_WATSC_DA

PHOSPHO_CORTEX_DA

PHOSPHO_SKMGN_DA

PHOSPHO_HEART_DA

PHOSPHO_KIDNEY_DA

PHOSPHO_LUNG_DA

PHOSPHO_LIVER_DA

PHOSPHO_WATSC_DA

ACETYL_HEART_DA

ACETYL_LIVER_DA

UBIQ_HEART_DA

UBIQ_LIVER_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on each
proteomics feature (protein, phosphosite, acetylsite, ubiquitlsite)
within each sex. One data frame per data type per tissue.
}
\keyword{datasets}
