% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REPEATED_FEATURES}
\alias{REPEATED_FEATURES}
\title{Repeated feature info}
\format{
A data frame with 825 rows and 10 variables:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{assay}}{character, assay abbreviation, one of \link{ASSAY_ABBREV}}
\item{\code{tissue}}{character, tissue abbreviation, one of \link{TISSUE_ABBREV}}
\item{\code{dataset}}{character, for immunoassay and metabolomics features, this variable specifies the immune panel (rat-myokine, rat-pituitary, rat-mag27plex, rat-metabolic, ADIPONECTIN, SERPIN-E) or metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features}
\item{\code{selection_fdr}}{double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite}
\item{\code{feature}}{character, duplicated \code{feature} in the format \code{\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};[feature_ID]}}
\item{\code{new_feature_ID}}{character, new unique \code{feature_ID} with \code{dataset} prepended}
\item{\code{new_feature}}{character, new unique \code{feature} in the format \code{\link{ASSAY_ABBREV};\link{TISSUE_ABBREV};[new_feature_ID]}}
}
}
\usage{
REPEATED_FEATURES
}
\description{
Documentation of 91 IMMUNO and METAB features that were measured on multiple platforms
and have multiple differential analysis results. New unique identifiers are defined. These new
unique identifiers were used in the graphical analysis.
}
\details{
91 IMMUNO and METAB features were measured on multiple platforms and have multiple differential analysis results.
In order to distinguish between the different sets of results, these
feature_IDs were prepended with \code{dataset}.
For example, "BDNF" is separated into "rat-myokine:BDNF" and "rat-pituitary:BDNF".
For completeness, both the modified and unmodified feature_IDs are included in both
the feature-to-gene map and universe lists. Note that the graphical analysis
uses the modified feature_IDs; differential analysis results use the unmodified feature_IDs.
}
\keyword{datasets}
