% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REPEATED_FEATURES}
\alias{REPEATED_FEATURES}
\title{Repeated feature info}
\format{
A data frame with 825 rows and 10 variables:
\describe{
\item{\code{feature_ID}}{character, MoTrPAC feature identifier}
\item{\code{assay_abbr}}{character, assay abbreviation}
\item{\code{tissue_abbreviation}}{character, tissue abbreviation}
\item{\code{dataset}}{character, platform for metabolomics features only.
"meta-reg" indicates the feature came from the merged meta-regression results.
See more details in \link[MotrpacBicQC:assay_codes]{MotrpacBicQC::assay_codes}.}
\item{\code{panel}}{character, name of LUMINEX panel for immunoassay features only}
\item{\code{selection_fdr}}{double, adjusted training p-value used to
select training-regulated analytes. P-values are IHW-adjusted across all
datasets within a given \code{assay} with \code{tissue} as a covariate.}
\item{\code{metabolite_refmet}}{character, RefMet name of metabolite OR the site-given code for unnamed metabolites}
\item{\code{feature}}{character, duplicated \code{feature} in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[feature_ID]}}
\item{\code{new_feature_ID}}{character, new unique \code{feature_ID} with \code{panel} or \code{dataset} prepended}
\item{\code{new_feature}}{character, new unique \code{feature} in the format \code{[ASSAY_ABBREV];[TISSUE_ABBREV];[new_feature_ID]}}
}
}
\usage{
REPEATED_FEATURES
}
\description{
Documentation of 91 IMMUNO and METAB features that were measured on multiple platforms
and have multiple differential analysis results. New unique identifiers are defined. These new
unique identifiers were used in the graphical analysis.
}
\details{
91 IMMUNO and METAB features were measured on multiple platforms and have multiple differential analysis results.
In order to distinguish between the different sets of results, these
feature_IDs were prepended with \code{panel} (for IMMUNO) or \code{dataset} (for METAB).
For example, "BDNF" is separated into "rat-myokine:BDNF" and "rat-pituitary:BDNF".
For completeness, both the modified and unmodified feature_IDs are included in both
the feature-to-gene map and universe lists. Note that the graphical analysis
uses the modified feature_IDs; differential analysis results use the unmodified feature_IDs.
}
\keyword{datasets}
