% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TISSUE_CODE_TO_ABBREV}
\alias{TISSUE_CODE_TO_ABBREV}
\title{Tissue code-to-abbreviation mapping}
\format{
Named character vector
}
\usage{
TISSUE_CODE_TO_ABBREV
}
\description{
Values are abbreviations \link{TISSUE_ABBREV} and names are codes.
See \link[MotrpacBicQC:bic_animal_tissue_code]{MotrpacBicQC::bic_animal_tissue_code} for more details.
}
\keyword{datasets}
