% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_DA}
\alias{TRNSCRPT_DA}
\alias{TRNSCRPT_BLOOD_DA}
\alias{TRNSCRPT_HIPPOC_DA}
\alias{TRNSCRPT_CORTEX_DA}
\alias{TRNSCRPT_HYPOTH_DA}
\alias{TRNSCRPT_SKMGN_DA}
\alias{TRNSCRPT_SKMVL_DA}
\alias{TRNSCRPT_HEART_DA}
\alias{TRNSCRPT_KIDNEY_DA}
\alias{TRNSCRPT_ADRNL_DA}
\alias{TRNSCRPT_COLON_DA}
\alias{TRNSCRPT_SPLEEN_DA}
\alias{TRNSCRPT_TESTES_DA}
\alias{TRNSCRPT_OVARY_DA}
\alias{TRNSCRPT_VENACV_DA}
\alias{TRNSCRPT_LUNG_DA}
\alias{TRNSCRPT_SMLINT_DA}
\alias{TRNSCRPT_LIVER_DA}
\alias{TRNSCRPT_BAT_DA}
\alias{TRNSCRPT_WATSC_DA}
\title{Differential analysis of RNA-seq datasets}
\format{
A data frame with 22 variables:
\describe{
\item{\code{feature}}{@eval feature()}
\item{\code{assay}}{@eval assay()}
\item{\code{assay_code}}{@eval assay_code()}
\item{\code{tissue}}{@eval tissue()}
\item{\code{tissue_code}}{@eval tissue_code()}
\item{\code{feature_ID}}{@eval feature_ID()}
\item{\code{sex}}{@eval sex()}
\item{\code{comparison_group}}{@eval comparison_group()}
\item{\code{p_value}}{@eval p_value()}
\item{\code{adj_p_value}}{@eval adj_p_value()}
\item{\code{logFC}}{@eval logFC()}
\item{\code{logFC_se}}{@eval logFC_se()}
\item{\code{shrunk_logFC}}{double, log fold-change with shrinkage applied}
\item{\code{shrunk_logFC_se}}{double, standard error of the shrunken log fold-change}
\item{\code{zscore}}{@eval zscore()}
\item{\code{covariates}}{@eval covariates()}
\item{\code{removed_samples}}{@eval removed_samples()}
\item{\code{comparison_average_intensity}}{@eval comparison_average_intensity()}
\item{\code{comparison_average_intensity_se}}{@eval comparison_average_intensity_se()}
\item{\code{reference_average_intensity}}{@eval reference_average_intensity()}
\item{\code{reference_average_intensity_se}}{@eval reference_average_intensity_se()}
\item{\code{selection_fdr}}{@eval selection_fdr()}
}

An object of class \code{data.frame} with 132864 rows and 22 columns.

An object of class \code{data.frame} with 131544 rows and 22 columns.

An object of class \code{data.frame} with 135640 rows and 22 columns.

An object of class \code{data.frame} with 110192 rows and 22 columns.

An object of class \code{data.frame} with 111784 rows and 22 columns.

An object of class \code{data.frame} with 115560 rows and 22 columns.

An object of class \code{data.frame} with 127848 rows and 22 columns.

An object of class \code{data.frame} with 126472 rows and 22 columns.

An object of class \code{data.frame} with 132448 rows and 22 columns.

An object of class \code{data.frame} with 130552 rows and 22 columns.

An object of class \code{data.frame} with 69436 rows and 22 columns.

An object of class \code{data.frame} with 68140 rows and 22 columns.

An object of class \code{data.frame} with 98028 rows and 22 columns.

An object of class \code{data.frame} with 132040 rows and 22 columns.

An object of class \code{data.frame} with 131288 rows and 22 columns.

An object of class \code{data.frame} with 115496 rows and 22 columns.

An object of class \code{data.frame} with 129232 rows and 22 columns.

An object of class \code{data.frame} with 134112 rows and 22 columns.
}
\usage{
TRNSCRPT_BLOOD_DA

TRNSCRPT_HIPPOC_DA

TRNSCRPT_CORTEX_DA

TRNSCRPT_HYPOTH_DA

TRNSCRPT_SKMGN_DA

TRNSCRPT_SKMVL_DA

TRNSCRPT_HEART_DA

TRNSCRPT_KIDNEY_DA

TRNSCRPT_ADRNL_DA

TRNSCRPT_COLON_DA

TRNSCRPT_SPLEEN_DA

TRNSCRPT_TESTES_DA

TRNSCRPT_OVARY_DA

TRNSCRPT_VENACV_DA

TRNSCRPT_LUNG_DA

TRNSCRPT_SMLINT_DA

TRNSCRPT_LIVER_DA

TRNSCRPT_BAT_DA

TRNSCRPT_WATSC_DA
}
\description{
Timewise summary statistics and training FDR from
differential analysis (DA) that tests the effect of training on the
expression of each gene within each sex. One data frame per tissue.
}
\keyword{datasets}
